refactor(hooks): split useCachedData.ts into focused modules
useCachedData.ts was 3136 lines long and contained unrelated concerns
mixed together. Split it into focused modules by domain:

- lib/cache/fetcherUtils.ts: internal API fetch helpers (fetchAPI,
  fetchFromAllClusters, fetchViaSSE, fetchViaGitOpsSSE, fetchGitOpsAPI,
  fetchRbacAPI, abortAllFetches)
- hooks/useCachedData/agentFetchers.ts: kc-agent HTTP fetchers
  (fetchPodIssuesViaAgent, fetchDeploymentsViaAgent, fetchWorkloadsFromAgent)
- hooks/useCachedData/demoData.ts: all synthetic demo data generators
- hooks/useCachedCoreWorkloads.ts: pods, events, pod-issues, deployment-issues,
  deployments, services, security issues, workloads, allPods hooks
- hooks/useCachedNodes.ts: node hooks (useCachedNodes, useCachedAllNodes,
  useCachedCoreDNSStatus) with CoreDNSClusterStatus type definitions
- hooks/useCachedGPU.ts: GPU nodes, GPU node health, GPU health CronJob,
  hardware health, warning events hooks; DeviceAlert/HardwareHealthData types
- hooks/useCachedK8sResources.ts: PVCs, namespaces, jobs, HPAs, configmaps,
  secrets, serviceaccounts, replicasets, statefulsets, daemonsets, cronjobs,
  ingresses, networkpolicies hooks
- hooks/useCachedGitOps.ts: Helm, operators, GitOps drifts, buildpacks, RBAC
  hooks (k8sRoles, k8sRoleBindings, k8sServiceAccounts)

useCachedData.ts is now a thin barrel that re-exports everything from the
focused modules plus the coreFetchers/specialtyFetchers prefetch helpers.
All 97 consumer files continue importing from useCachedData without changes.

Fixes #8624
