+++ web/src/hooks/useCachedData.ts
@@ -13,981 +13,105 @@
  * - etc.
  *
  * The hooks maintain the same return interface for easy migration.
+ *
+ * ---------------------------------------------------------------------------
+ * Module layout (extracted for maintainability — issue #8624):
+ *
+ *   lib/cache/fetcherUtils.ts   — internal API fetch helpers (fetchAPI, SSE, etc.)
+ *   hooks/useCachedData/
+ *     agentFetchers.ts          — kc-agent HTTP fetchers
+ *     demoData.ts               — synthetic demo data generators
+ *   hooks/useCachedCoreWorkloads.ts  — pods, events, deployments, security, workloads
+ *   hooks/useCachedNodes.ts          — node hooks + CoreDNS
+ *   hooks/useCachedGPU.ts            — GPU nodes, GPU health, hardware health, warning events
+ *   hooks/useCachedK8sResources.ts   — PVCs, namespaces, jobs, HPAs, configmaps, …
+ *   hooks/useCachedGitOps.ts         — Helm, operators, GitOps drift, buildpacks, RBAC
+ *   hooks/useCachedProw.ts           — Prow CI (pre-existing)
+ *   hooks/useCachedLLMd.ts           — LLM-d (pre-existing)
+ *   hooks/useCachedISO27001.ts       — ISO 27001 audit (pre-existing)
+ * ---------------------------------------------------------------------------
  */
-
-import { useState } from 'react'
-import { useCache, type RefreshCategory } from '../lib/cache'
-import { isBackendUnavailable, authFetch } from '../lib/api'
-import { kubectlProxy } from '../lib/kubectlProxy'
-import { fetchSSE } from '../lib/sseClient'
-import { clusterCacheRef, deduplicateClustersByServer } from './mcp/shared'
-import { isAgentUnavailable } from './useLocalAgent'
-import { LOCAL_AGENT_HTTP_URL, STORAGE_KEY_TOKEN } from '../lib/constants'
-import { FETCH_DEFAULT_TIMEOUT_MS, AI_PREDICTION_TIMEOUT_MS, KUBECTL_EXTENDED_TIMEOUT_MS } from '../lib/constants/network'
-import { settledWithConcurrency } from '../lib/utils/concurrency'
-import type {
-  PodInfo,
-  PodIssue,
-  ClusterEvent,
-  DeploymentIssue,
-  Deployment,
-  Service,
-  SecurityIssue,
-  NodeInfo,
-  GPUNode,
-  GPUNodeHealthStatus,
-  PVC,
-  Job,
-  HPA,
-  HelmRelease,
-  HelmHistoryEntry,
-  Operator,
-  OperatorSubscription,
-  GitOpsDrift,
-  BuildpackImage,
-  ConfigMap,
-  Secret,
-  ServiceAccount,
-  ReplicaSet,
-  StatefulSet,
-  DaemonSet,
-  CronJob,
-  Ingress,
-  NetworkPolicy,
-  K8sRole,
-  K8sRoleBinding,
-  K8sServiceAccountInfo,
-  GPUHealthCronJobStatus } from './useMCP'
-import type { Workload } from './useWorkloads'
-import { fetchProwJobs } from './useCachedProw'
-import { fetchLLMdServers, fetchLLMdModels } from './useCachedLLMd'
-import {
-  SecurityIssuesResponseSchema,
-  PodsResponseSchema,
-  EventsResponseSchema,
-  DeploymentsResponseSchema,
-  NodesResponseSchema,
-  GPUNodesResponseSchema,
-  GPUNodeHealthResponseSchema,
-  ClustersResponseSchema,
-} from '../lib/schemas'
-import { validateResponse, validateArrayResponse } from '../lib/schemas/validate'
 
 // ============================================================================
-// API Fetchers
+// Public API surface — re-export everything from focused modules
 // ============================================================================
 
-const getToken = () => localStorage.getItem(STORAGE_KEY_TOKEN)
-
-const AGENT_HTTP_TIMEOUT_MS = 5_000
-
-/** Maximum number of pods to return from a prefetch query */
-const MAX_PREFETCH_PODS = 100
-
-// Global AbortController for all in-flight fetchAPI requests.
-// Aborting this cancels every pending cluster fetch, freeing browser
-// connections instantly so navigation and lazy chunks can load.
-let globalFetchController = new AbortController()
-
-/** Abort all in-flight fetchAPI requests (e.g. on route change). */
-export function abortAllFetches(): void {
-  globalFetchController.abort()
-  globalFetchController = new AbortController()
-}
-
-async function fetchAPI<T>(
-  endpoint: string,
-  params?: Record<string, string | number | undefined>
-): Promise<T> {
-  const token = getToken()
-  if (!token) {
-    throw new Error('No authentication token')
-  }
-
-  const searchParams = new URLSearchParams()
-  if (params) {
-    Object.entries(params).forEach(([key, value]) => {
-      if (value !== undefined) {
-        searchParams.append(key, String(value))
-      }
-    })
-  }
-
-  const url = `${LOCAL_AGENT_HTTP_URL}/${endpoint}?${searchParams}`
-  // Combine the global abort signal with a per-request timeout.
-  // AbortSignal.any() fires if either signal triggers.
-  const signal = AbortSignal.any([
-    globalFetchController.signal,
-    AbortSignal.timeout(FETCH_DEFAULT_TIMEOUT_MS),
-  ])
-  const response = await fetch(url, {
-    method: 'GET',
-    headers: {
-      'Content-Type': 'application/json',
-      Authorization: `Bearer ${token}` },
-    signal })
-
-  if (!response.ok) {
-    throw new Error(`API error: ${response.status}`)
-  }
-
-  const text = await response.text()
-  try {
-    return JSON.parse(text) as T
-  } catch {
-    throw new Error(`API returned non-JSON response from /api/mcp/${endpoint}`)
-  }
-}
-
-// Get list of reachable (or not-yet-checked) clusters (prefer local agent data for accurate reachability)
-function getReachableClusters(): string[] {
-  // Use local agent's cluster cache - it has up-to-date reachability info
-  // Include reachable === undefined (health check pending) — same logic as getAgentClusters
-  if (clusterCacheRef.clusters.length > 0) {
-    return clusterCacheRef.clusters
-      .filter(c => c.reachable !== false && !c.name.includes('/'))
-      .map(c => c.name)
-  }
-  return []
-}
-
-// Fetch list of available clusters from backend (fallback)
-async function fetchClusters(): Promise<string[]> {
-  // First check local agent data - faster and more accurate reachability
-  const localClusters = getReachableClusters()
-  if (localClusters.length > 0) {
-    return localClusters
-  }
-
-  // Fall back to backend API
-  const raw = await fetchAPI<unknown>('clusters')
-  const data = validateArrayResponse<{ clusters: Array<{ name: string; reachable?: boolean }> }>(ClustersResponseSchema, raw, '/api/mcp/clusters', 'clusters')
-  return (data.clusters || [])
-    .filter(c => c.reachable !== false && !c.name.includes('/'))
-    .map(c => c.name)
-}
-
-/**
- * Options for fetchFromAllClusters that alter empty-result semantics.
- */
-interface FetchFromAllClustersOptions {
-  /**
-   * When true, throw if any cluster fetch failed AND the accumulated result
-   * is empty. Intended for endpoints where an empty result from a partially
-   * failed fan-out is ambiguous (we cannot tell "truly zero" from "the one
-   * cluster with data errored"). Forces callers into the error path so
-   * existing cached data is preserved. See issues #8080, #8081 (GPU flap).
-   */
-  throwIfPartialFailureEmpty?: boolean
-}
-
-// Fetch data from all clusters in parallel and merge results
-// Throws if ALL cluster fetches fail (so callers can fall back to agent)
-async function fetchFromAllClusters<T>(
-  endpoint: string,
-  resultKey: string,
-  params?: Record<string, string | number | undefined>,
-  addClusterField = true,
-  onProgress?: (partial: T[]) => void,
-  options?: FetchFromAllClustersOptions
-): Promise<T[]> {
-  const clusters = await fetchClusters()
-  if (clusters.length === 0) {
-    throw new Error('No clusters available (agent connecting or backend not authenticated)')
-  }
-
-  // (#6857) Each callback returns its own tagged items instead of pushing
-  // into a shared accumulator. Results are aggregated after all tasks settle,
-  // eliminating the shared-mutation hazard across concurrent await points.
-  const tasks = clusters.map((cluster) => async () => {
-    const data = await fetchAPI<Record<string, T[]>>(endpoint, { ...params, cluster })
-    const items = data[resultKey] || []
-    return addClusterField ? items.map(item => ({ ...item, cluster })) : items
-  })
-
-  // Use onSettled callback to push partial results to the UI as each
-  // cluster responds, instead of waiting for all clusters (including
-  // unreachable ones with long timeouts) to complete.
-  const accumulated: T[] = []
-  let failedCount = 0
-
-  // Named handler — onSettled runs sequentially (not concurrent with tasks),
-  // so mutating accumulated/failedCount here is safe.  Declared as a function
-  // so the concurrent-mutation-safety static scan only analyses task callbacks.
-  function handleSettled(result: PromiseSettledResult<T[]>) {
-    if (result.status === 'fulfilled') {
-      accumulated.push(...result.value)
-      onProgress?.([...accumulated])
-    } else {
-      failedCount++
-    }
-  }
-  await settledWithConcurrency(tasks, undefined, handleSettled)
-
-  // If every cluster fetch failed, throw so callers can try agent fallback
-  if (accumulated.length === 0 && clusters.length > 0 && failedCount === clusters.length) {
-    throw new Error('All cluster fetches failed')
-  }
-
-  // Opt-in partial-failure protection: if any cluster failed and the
-  // accumulated result is empty, the empty result is ambiguous — we cannot
-  // distinguish "no data anywhere" from "the one cluster with data errored".
-  // Callers that set throwIfPartialFailureEmpty=true prefer preserving stale
-  // cache via the error path over silently overwriting it with []. See
-  // issues #8080, #8081 where a transient fetch failure for the vllm-d
-  // cluster caused GPU Usage Trend to flap to "No GPU Nodes".
-  if (
-    options?.throwIfPartialFailureEmpty &&
-    accumulated.length === 0 &&
-    failedCount > 0
-  ) {
-    throw new Error(
-      `Partial cluster failure yielded empty result (${failedCount}/${clusters.length} clusters errored) — preserving existing cache`,
-    )
-  }
-
-  return accumulated
-}
-
-/**
- * Fetch data from all clusters using SSE streaming.
- * Each cluster's data arrives as a separate event, allowing progressive rendering.
- * Falls back to fetchFromAllClusters if SSE fails or is unavailable.
- */
-async function fetchViaSSE<T>(
-  endpoint: string,
-  resultKey: string,
-  params?: Record<string, string | number | undefined>,
-  onProgress?: (partial: T[]) => void,
-  options?: FetchFromAllClustersOptions
-): Promise<T[]> {
-  const token = getToken()
-  // SSE only available with real backend token
-  if (!token || token === 'demo-token' || isBackendUnavailable()) {
-    return fetchFromAllClusters<T>(endpoint, resultKey, params, true, onProgress, options)
-  }
-
-  try {
-    const accumulated: T[] = []
-    // Track backend-emitted cluster_error events so we can detect the
-    // "partial failure yielding empty result" case below (issues #8080,
-    // #8081). Without this, a cluster_error on the only GPU-bearing
-    // cluster would silently resolve to [] and flap the UI.
-    let clusterErrorCount = 0
-    const result = await fetchSSE<T>({
-      url: `${LOCAL_AGENT_HTTP_URL}/${endpoint}/stream`,
-      params,
-      itemsKey: resultKey,
-      onClusterData: (_cluster, items) => {
-        accumulated.push(...items)
-        onProgress?.([...accumulated])
-      },
-      onClusterError: () => {
-        clusterErrorCount += 1
-      } })
-    if (
-      options?.throwIfPartialFailureEmpty &&
-      result.length === 0 &&
-      clusterErrorCount > 0
-    ) {
-      throw new Error(
-        `Partial SSE failure yielded empty result (${clusterErrorCount} cluster_error events) — preserving existing cache`,
-      )
-    }
-    return result
-  } catch {
-    // SSE failed — fall back to per-cluster REST
-    return fetchFromAllClusters<T>(endpoint, resultKey, params, true, onProgress, options)
-  }
-}
-
-/**
- * Fetch GitOps data via SSE streaming (uses /api/gitops/ prefix).
- */
-async function fetchViaGitOpsSSE<T>(
-  endpoint: string,
-  resultKey: string,
-  params?: Record<string, string | number | undefined>,
-  onProgress?: (partial: T[]) => void
-): Promise<T[]> {
-  const token = getToken()
-  if (!token || token === 'demo-token' || isBackendUnavailable()) {
-    throw new Error('No data source available (backend not authenticated)')
-  }
-
-  const accumulated: T[] = []
-  return await fetchSSE<T>({
-    url: `/api/gitops/${endpoint}/stream`,
-    params,
-    itemsKey: resultKey,
-    onClusterData: (_cluster, items) => {
-      accumulated.push(...items)
-      onProgress?.([...accumulated])
-    } })
-}
-
-/**
- * Fetch data from a GitOps REST endpoint.
- */
-async function fetchGitOpsAPI<T>(
-  endpoint: string,
-  params?: Record<string, string | number | undefined>
-): Promise<T> {
-  const token = getToken()
-  if (!token) throw new Error('No authentication token')
-
-  const searchParams = new URLSearchParams()
-  if (params) {
-    Object.entries(params).forEach(([key, value]) => {
-      if (value !== undefined) searchParams.append(key, String(value))
-    })
-  }
-
-  const url = `/api/gitops/${endpoint}?${searchParams}`
-  const response = await fetch(url, {
-    method: 'GET',
-    headers: { 'Content-Type': 'application/json', Authorization: `Bearer ${token}` },
-    signal: AbortSignal.timeout(FETCH_DEFAULT_TIMEOUT_MS) })
-
-  if (!response.ok) throw new Error(`API error: ${response.status}`)
-  const gitopsText = await response.text()
-  try {
-    return JSON.parse(gitopsText) as T
-  } catch {
-    throw new Error(`API returned non-JSON response from /api/gitops/${endpoint}`)
-  }
-}
-
-/** RBAC timeout — roles/bindings can be slow on large clusters */
-const RBAC_FETCH_TIMEOUT_MS = 60_000
-
-/**
- * Fetch data from an RBAC REST endpoint (/api/rbac/).
- */
-async function fetchRbacAPI<T>(
-  endpoint: string,
-  params?: Record<string, string | number | boolean | undefined>
-): Promise<T> {
-  const token = getToken()
-  if (!token) throw new Error('No authentication token')
-
-  const searchParams = new URLSearchParams()
-  if (params) {
-    Object.entries(params).forEach(([key, value]) => {
-      if (value !== undefined) searchParams.append(key, String(value))
-    })
-  }
-
-  const url = `/api/rbac/${endpoint}?${searchParams}`
-  const response = await fetch(url, {
-    method: 'GET',
-    headers: { 'Content-Type': 'application/json', Authorization: `Bearer ${token}` },
-    signal: AbortSignal.timeout(RBAC_FETCH_TIMEOUT_MS) })
-
-  if (!response.ok) throw new Error(`API error: ${response.status}`)
-  const rbacText = await response.text()
-  try {
-    return JSON.parse(rbacText) as T
-  } catch {
-    throw new Error(`API returned non-JSON response from /api/rbac/${endpoint}`)
-  }
-}
-
-// ============================================================================
-// Agent-based fetchers (used when backend is unavailable but agent is connected)
-// ============================================================================
-
-/** Get reachable (or not-yet-checked) cluster names from the shared cluster cache (deduplicated) */
-function getAgentClusters(): Array<{ name: string; context?: string }> {
-  // useCache prevents calling fetchers in demo mode via effectiveEnabled
-  // Skip long context-path names (contain '/') — these are duplicates of short-named aliases
-  // e.g. "default/api-fmaas-vllm-d-...:6443/..." duplicates "vllm-d"
-  // Include clusters with reachable === undefined (health check pending) to avoid
-  // race condition where cards fetch before health checks complete and cache empty results.
-  // This matches the backend's HealthyClusters() which treats unknown as healthy.
-  return clusterCacheRef.clusters
-    .filter(c => c.reachable !== false && !c.name.includes('/'))
-    .map(c => ({ name: c.name, context: c.context }))
-}
-
-/** Fetch pod issues from all clusters via agent kubectl proxy */
-async function fetchPodIssuesViaAgent(namespace?: string, onProgress?: (partial: PodIssue[]) => void): Promise<PodIssue[]> {
-  if (isAgentUnavailable()) return []
-  const clusters = getAgentClusters()
-  if (clusters.length === 0) return []
-
-  const tasks = clusters.map(({ name, context }) => async () => {
-    const ctx = context || name
-    const issues = await kubectlProxy.getPodIssues(ctx, namespace)
-    // Always use the short name — kubectlProxy returns context path as cluster
-    return issues.map(i => ({ ...i, cluster: name }))
-  })
-
-  const accumulated: PodIssue[] = []
-  function handleSettled(result: PromiseSettledResult<PodIssue[]>) {
-    if (result.status === 'fulfilled') {
-      accumulated.push(...result.value)
-      onProgress?.([...accumulated])
-    }
-  }
-  await settledWithConcurrency(tasks, undefined, handleSettled)
-  return accumulated
-}
-
-/** Fetch deployments from all clusters via agent HTTP endpoint */
-async function fetchDeploymentsViaAgent(namespace?: string, onProgress?: (partial: Deployment[]) => void): Promise<Deployment[]> {
-  if (isAgentUnavailable()) return []
-  const clusters = getAgentClusters()
-  if (clusters.length === 0) return []
-
-  const tasks = clusters.map(({ name, context }) => async () => {
-    const params = new URLSearchParams()
-    params.append('cluster', context || name)
-    if (namespace) params.append('namespace', namespace)
-
-    const controller = new AbortController()
-    const timeoutId = setTimeout(() => controller.abort(), AGENT_HTTP_TIMEOUT_MS)
-    const response = await fetch(`${LOCAL_AGENT_HTTP_URL}/deployments?${params}`, {
-      signal: controller.signal,
-      headers: { Accept: 'application/json' } })
-    clearTimeout(timeoutId)
-
-    if (!response.ok) throw new Error(`Agent returned ${response.status}`)
-    // Use .catch() on .json() to prevent Firefox from firing unhandledrejection
-    // before the outer try/catch processes the rejection (microtask timing issue).
-    const data = await response.json().catch(() => null)
-    if (!data) throw new Error('Invalid JSON response from agent')
-    // Always use the short name — agent echoes back context path as cluster
-    return ((data.deployments || []) as Deployment[]).map(d => ({
-      ...d,
-      cluster: name }))
-  })
-
-  const accumulated: Deployment[] = []
-  function handleSettled(result: PromiseSettledResult<Deployment[]>) {
-    if (result.status === 'fulfilled') {
-      accumulated.push(...result.value)
-      onProgress?.([...accumulated])
-    }
-  }
-  await settledWithConcurrency(tasks, undefined, handleSettled)
-  return accumulated
-}
-
-// ============================================================================
-// Demo Data (fallbacks)
-// ============================================================================
-
-const getDemoPods = (): PodInfo[] => [
-  { name: 'frontend-7d8f9c4b5-x2km4', namespace: 'production', status: 'Running', ready: '1/1', restarts: 0, age: '2d', cpuRequestMillis: 500, memoryRequestBytes: 536870912, cpuUsageMillis: 320, memoryUsageBytes: 412516352, metricsAvailable: true },
-  { name: 'backend-api-6c8d7f5e4-j3ln9', namespace: 'production', status: 'Running', ready: '2/2', restarts: 1, age: '5d', cpuRequestMillis: 1000, memoryRequestBytes: 1073741824, cpuUsageMillis: 850, memoryUsageBytes: 892871680, metricsAvailable: true },
-  { name: 'ml-worker-8f9a6b7c3-k4lm2', namespace: 'ml-workloads', status: 'Running', ready: '1/1', restarts: 0, age: '1d', cpuRequestMillis: 4000, memoryRequestBytes: 8589934592, gpuRequest: 2, cpuUsageMillis: 3200, memoryUsageBytes: 7516192768, metricsAvailable: true },
-  { name: 'inference-server-5d4c3b2a1-n7op9', namespace: 'ml-workloads', status: 'Running', ready: '1/1', restarts: 2, age: '3d', cpuRequestMillis: 2000, memoryRequestBytes: 4294967296, gpuRequest: 1, cpuUsageMillis: 1850, memoryUsageBytes: 3865470566, metricsAvailable: true },
-  { name: 'cache-redis-6e5d4c3b2-q8rs1', namespace: 'production', status: 'Running', ready: '1/1', restarts: 0, age: '7d', cpuRequestMillis: 250, memoryRequestBytes: 268435456, cpuUsageMillis: 45, memoryUsageBytes: 134217728, metricsAvailable: true },
-]
-
-const getDemoEvents = (): ClusterEvent[] => {
-  const now = Date.now()
-  const minutesAgo = (m: number) => new Date(now - m * 60000).toISOString()
-  return [
-    { type: 'Warning', reason: 'FailedScheduling', message: 'No nodes available to schedule pod', object: 'Pod/worker-5c6d7e8f9-n3p2q', namespace: 'default', cluster: 'eks-prod-us-east-1', count: 3, firstSeen: minutesAgo(25), lastSeen: minutesAgo(5) },
-    { type: 'Normal', reason: 'Started', message: 'Container started successfully', object: 'Pod/web-frontend-8e9f0a1b2-def34', namespace: 'production', cluster: 'gke-staging', count: 1, firstSeen: minutesAgo(12), lastSeen: minutesAgo(12) },
-    { type: 'Warning', reason: 'BackOff', message: 'Back-off restarting failed container', object: 'Pod/api-server-7d8f9c6b5-x2k4m', namespace: 'production', cluster: 'eks-prod-us-east-1', count: 15, firstSeen: minutesAgo(45), lastSeen: minutesAgo(2) },
-    { type: 'Normal', reason: 'Pulled', message: 'Container image pulled successfully', object: 'Pod/frontend-8e9f0a1b2-def34', namespace: 'production', cluster: 'gke-staging', count: 1, firstSeen: minutesAgo(8), lastSeen: minutesAgo(8) },
-    { type: 'Warning', reason: 'Unhealthy', message: 'Readiness probe failed: connection refused', object: 'Pod/cache-redis-0', namespace: 'data', cluster: 'gke-staging', count: 8, firstSeen: minutesAgo(30), lastSeen: minutesAgo(1) },
-    { type: 'Normal', reason: 'ScalingReplicaSet', message: 'Scaled up replica set api-gateway-7d8c to 3', object: 'Deployment/api-gateway', namespace: 'production', cluster: 'eks-prod-us-east-1', count: 1, firstSeen: minutesAgo(18), lastSeen: minutesAgo(18) },
-    { type: 'Normal', reason: 'SuccessfulCreate', message: 'Created pod: worker-5c6d7e8f9-abc12', object: 'ReplicaSet/worker-5c6d7e8f9', namespace: 'batch', cluster: 'vllm-gpu-cluster', count: 1, firstSeen: minutesAgo(22), lastSeen: minutesAgo(22) },
-    { type: 'Warning', reason: 'FailedMount', message: 'MountVolume.SetUp failed for volume "config": configmap "app-config" not found', object: 'Pod/ml-inference-7f8g9h-xyz99', namespace: 'ml', cluster: 'vllm-gpu-cluster', count: 4, firstSeen: minutesAgo(35), lastSeen: minutesAgo(3) },
-  ]
-}
-
-const getDemoPodIssues = (): PodIssue[] => [
-  { name: 'api-server-7d8f9c6b5-x2k4m', namespace: 'production', cluster: 'eks-prod-us-east-1', status: 'CrashLoopBackOff', issues: ['Container restarting', 'OOMKilled'], restarts: 15 },
-  { name: 'worker-5c6d7e8f9-n3p2q', namespace: 'batch', cluster: 'vllm-gpu-cluster', status: 'ImagePullBackOff', issues: ['Failed to pull image'], restarts: 0 },
-  { name: 'cache-redis-0', namespace: 'data', cluster: 'gke-staging', status: 'Pending', issues: ['Insufficient memory'], restarts: 0 },
-  { name: 'metrics-collector-2b4c6-j8k9l', namespace: 'monitoring', cluster: 'aks-dev-westeu', status: 'CrashLoopBackOff', issues: ['Exit code 137'], restarts: 8 },
-  { name: 'gpu-scheduler-0', namespace: 'ml-ops', cluster: 'vllm-gpu-cluster', status: 'Pending', issues: ['Insufficient nvidia.com/gpu'], restarts: 0 },
-]
-
-const getDemoDeploymentIssues = (): DeploymentIssue[] => [
-  { name: 'web-frontend', namespace: 'production', replicas: 3, readyReplicas: 2, reason: 'ReplicaFailure' },
-]
-
-const getDemoDeployments = (): Deployment[] => [
-  { name: 'web-frontend', namespace: 'production', cluster: 'eks-prod-us-east-1', status: 'running', replicas: 3, readyReplicas: 3, updatedReplicas: 3, availableReplicas: 3, progress: 100 },
-  { name: 'api-gateway', namespace: 'production', cluster: 'eks-prod-us-east-1', status: 'deploying', replicas: 3, readyReplicas: 1, updatedReplicas: 2, availableReplicas: 1, progress: 33 },
-  { name: 'worker-service', namespace: 'batch', cluster: 'gke-staging', status: 'deploying', replicas: 4, readyReplicas: 2, updatedReplicas: 3, availableReplicas: 2, progress: 50 },
-  { name: 'ml-inference', namespace: 'ml', cluster: 'vllm-gpu-cluster', status: 'deploying', replicas: 2, readyReplicas: 0, updatedReplicas: 1, availableReplicas: 0, progress: 0 },
-  { name: 'cache-redis', namespace: 'data', cluster: 'gke-staging', status: 'running', replicas: 3, readyReplicas: 3, updatedReplicas: 3, availableReplicas: 3, progress: 100 },
-  { name: 'monitoring-stack', namespace: 'monitoring', cluster: 'aks-dev-westeu', status: 'running', replicas: 2, readyReplicas: 2, updatedReplicas: 2, availableReplicas: 2, progress: 100 },
-]
-
-const getDemoServices = (): Service[] => [
-  { name: 'web-service', namespace: 'production', type: 'LoadBalancer', clusterIP: '10.0.0.1', ports: ['80/TCP'] },
-]
-
-const getDemoSecurityIssues = (): SecurityIssue[] => [
-  { name: 'api-server-7d8f9c6b5-x2k4m', namespace: 'production', cluster: 'eks-prod-us-east-1', issue: 'Privileged container', severity: 'high', details: 'Container running in privileged mode' },
-  { name: 'worker-deployment', namespace: 'batch', cluster: 'vllm-gpu-cluster', issue: 'Running as root', severity: 'high', details: 'Container running as root user' },
-  { name: 'nginx-ingress', namespace: 'ingress', cluster: 'eks-prod-us-east-1', issue: 'Host network enabled', severity: 'medium', details: 'Pod using host network namespace' },
-  { name: 'monitoring-agent', namespace: 'monitoring', cluster: 'gke-staging', issue: 'Missing security context', severity: 'low', details: 'No security context defined' },
-  { name: 'redis-cache', namespace: 'data', cluster: 'openshift-prod', issue: 'Capabilities not dropped', severity: 'medium', details: 'Container not dropping all capabilities' },
-  { name: 'legacy-app', namespace: 'legacy', cluster: 'vllm-gpu-cluster', issue: 'Running as root', severity: 'high', details: 'Container running as root user' },
-]
-
-// ============================================================================
-// Cached Data Hooks
-// ============================================================================
-
-interface CachedHookResult<T> {
-  data: T
-  isLoading: boolean
-  isRefreshing: boolean
-  isDemoFallback: boolean
-  error: string | null
-  isFailed: boolean
-  consecutiveFailures: number
-  lastRefresh: number | null
-  refetch: () => Promise<void>
-}
-
-/**
- * Hook for fetching pods with caching
- * When no cluster is specified, fetches from all available clusters
- */
-export function useCachedPods(
-  cluster?: string,
-  namespace?: string,
-  options?: { limit?: number; category?: RefreshCategory }
-): CachedHookResult<PodInfo[]> & { pods: PodInfo[] } {
-  const { limit = 100, category = 'pods' } = options || {}
-  const key = `pods:${cluster || 'all'}:${namespace || 'all'}:${limit}`
-
-  // Note: useCache handles demo mode detection internally via useSyncExternalStore
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as PodInfo[],
-    demoData: getDemoPods(),
-    fetcher: async () => {
-      let pods: PodInfo[]
-      if (cluster) {
-        const raw = await fetchAPI<unknown>('pods', { cluster, namespace })
-        const data = validateArrayResponse<{ pods: PodInfo[] }>(PodsResponseSchema, raw, '/api/mcp/pods', 'pods')
-        pods = (data.pods || []).map(p => ({ ...p, cluster }))
-      } else {
-        pods = await fetchFromAllClusters<PodInfo>('pods', 'pods', { namespace })
-      }
-      return pods
-        .sort((a, b) => (b.restarts || 0) - (a.restarts || 0))
-        .slice(0, limit)
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      const pods = await fetchViaSSE<PodInfo>('pods', 'pods', { namespace }, (partial) => {
-        onProgress(partial.sort((a, b) => (b.restarts || 0) - (a.restarts || 0)).slice(0, limit))
-      })
-      return pods
-        .sort((a, b) => (b.restarts || 0) - (a.restarts || 0))
-        .slice(0, limit)
-    } })
-
-  return {
-    pods: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching events with caching
- */
-export function useCachedEvents(
-  cluster?: string,
-  namespace?: string,
-  options?: { limit?: number; category?: RefreshCategory }
-): CachedHookResult<ClusterEvent[]> & { events: ClusterEvent[] } {
-  const { limit = 20, category = 'realtime' } = options || {}
-  const key = `events:${cluster || 'all'}:${namespace || 'all'}:${limit}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as ClusterEvent[],
-    demoData: getDemoEvents(),
-    fetcher: async () => {
-      // Try agent first (direct kubectl proxy — works before backend auth)
-      if (clusterCacheRef.clusters.length > 0 && !isAgentUnavailable()) {
-        if (cluster) {
-          const ci = clusterCacheRef.clusters.find(c => c.name === cluster)
-          const ctx = ci?.context || cluster
-          const events = await kubectlProxy.getEvents(ctx, namespace, limit)
-          return events.map(e => ({ ...e, cluster }))
-        }
-        // Fetch from all clusters via agent with bounded concurrency
-        const clusters = getAgentClusters()
-        const allEvents: ClusterEvent[] = []
-        const results = await settledWithConcurrency(
-          clusters.map((ci) => async () => {
-            const ctx = ci.context || ci.name
-            const events = await kubectlProxy.getEvents(ctx, namespace, limit)
-            return events.map(e => ({ ...e, cluster: ci.name }))
-          })
-        )
-        for (const r of (results || [])) {
-          if (r.status === 'fulfilled') allEvents.push(...r.value)
-        }
-        return allEvents
-          .sort((a, b) => {
-            const timeA = a.lastSeen ? new Date(a.lastSeen).getTime() : 0
-            const timeB = b.lastSeen ? new Date(b.lastSeen).getTime() : 0
-            return timeB - timeA
-          })
-          .slice(0, limit)
-      }
-
-      // Fall back to REST API (requires backend auth)
-      if (cluster) {
-        const raw = await fetchAPI<unknown>('events', { cluster, namespace, limit })
-        const data = validateArrayResponse<{ events: ClusterEvent[] }>(EventsResponseSchema, raw, '/api/mcp/events', 'events')
-        return data.events || []
-      }
-      return await fetchFromAllClusters<ClusterEvent>('events', 'events', { namespace, limit })
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      // Try agent-based progressive fetch first
-      if (clusterCacheRef.clusters.length > 0 && !isAgentUnavailable()) {
-        const clusters = getAgentClusters()
-        const accumulated: ClusterEvent[] = []
-        const tasks = (clusters || []).map((ci) => async () => {
-          const ctx = ci.context || ci.name
-          const events = await kubectlProxy.getEvents(ctx, namespace, limit)
-          return events.map(e => ({ ...e, cluster: ci.name }))
-        })
-        function handleSettled(result: PromiseSettledResult<ClusterEvent[]>) {
-          if (result.status === 'fulfilled') {
-            accumulated.push(...result.value)
-            accumulated.sort((a, b) => {
-              const timeA = a.lastSeen ? new Date(a.lastSeen).getTime() : 0
-              const timeB = b.lastSeen ? new Date(b.lastSeen).getTime() : 0
-              return timeB - timeA
-            })
-            onProgress([...accumulated].slice(0, limit))
-          }
-        }
-        await settledWithConcurrency(tasks, undefined, handleSettled)
-        return accumulated.slice(0, limit)
-      }
-      // Fall back to SSE via backend
-      return await fetchViaSSE<ClusterEvent>('events', 'events', { namespace, limit }, onProgress)
-    } })
-
-  return {
-    events: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching pod issues with caching
- * When no cluster is specified, fetches from all available clusters
- */
-export function useCachedPodIssues(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<PodIssue[]> & { issues: PodIssue[] } {
-  const { category = 'pods' } = options || {}
-  const key = `podIssues:${cluster || 'all'}:${namespace || 'all'}`
-
-  const sortIssues = (items: PodIssue[]) => items.sort((a, b) => (b.restarts || 0) - (a.restarts || 0))
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as PodIssue[],
-    demoData: getDemoPodIssues(),
-    fetcher: async () => {
-      let issues: PodIssue[]
-
-      // Try agent first (fast, no backend needed)
-      if (clusterCacheRef.clusters.length > 0 && !isAgentUnavailable()) {
-        if (cluster) {
-          const clusterInfo = clusterCacheRef.clusters.find(c => c.name === cluster)
-          const ctx = clusterInfo?.context || cluster
-          issues = await kubectlProxy.getPodIssues(ctx, namespace)
-          issues = issues.map(i => ({ ...i, cluster: cluster }))
-        } else {
-          issues = await fetchPodIssuesViaAgent(namespace)
-        }
-        return sortIssues(issues)
-      }
-
-      // Fall back to REST API
-      const token = getToken()
-      const hasRealToken = token && token !== 'demo-token'
-      if (hasRealToken && !isBackendUnavailable()) {
-        if (cluster) {
-          const data = await fetchAPI<{ issues: PodIssue[] }>('pod-issues', { cluster, namespace })
-          issues = (data.issues || []).map(i => ({ ...i, cluster }))
-        } else {
-          issues = await fetchFromAllClusters<PodIssue>('pod-issues', 'issues', { namespace })
-        }
-        return sortIssues(issues)
-      }
-
-      // No data source available yet — throw so cache preserves existing data and retries
-      throw new Error('No data source available (agent connecting or backend not authenticated)')
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      // Try agent first
-      if (clusterCacheRef.clusters.length > 0 && !isAgentUnavailable()) {
-        const issues = await fetchPodIssuesViaAgent(namespace, (partial) => {
-          onProgress(sortIssues([...partial]))
-        })
-        return sortIssues(issues)
-      }
-
-      // Fall back to SSE streaming -> REST per-cluster
-      const issues = await fetchViaSSE<PodIssue>('pod-issues', 'issues', { namespace }, (partial) => {
-        onProgress(sortIssues([...partial]))
-      })
-      return sortIssues(issues)
-    } })
-
-  return {
-    issues: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching deployment issues with caching
- */
-export function useCachedDeploymentIssues(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<DeploymentIssue[]> & { issues: DeploymentIssue[] } {
-  const { category = 'deployments' } = options || {}
-  const key = `deploymentIssues:${cluster || 'all'}:${namespace || 'all'}`
-
-  const deriveIssues = (deployments: Deployment[]): DeploymentIssue[] =>
-    deployments
-      .filter(d => (d.readyReplicas ?? 0) < (d.replicas ?? 1))
-      .map(d => ({
-        name: d.name,
-        namespace: d.namespace || 'default',
-        cluster: d.cluster,
-        replicas: d.replicas ?? 1,
-        readyReplicas: d.readyReplicas ?? 0,
-        reason: d.status === 'failed' ? 'DeploymentFailed' : 'ReplicaFailure' }))
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as DeploymentIssue[],
-    demoData: getDemoDeploymentIssues(),
-    fetcher: async () => {
-      // Try agent first — derive deployment issues from deployment data
-      if (clusterCacheRef.clusters.length > 0 && !isAgentUnavailable()) {
-        const deployments = cluster
-          ? await (async () => {
-              const clusterInfo = clusterCacheRef.clusters.find(c => c.name === cluster)
-              const params = new URLSearchParams()
-              params.append('cluster', clusterInfo?.context || cluster)
-              if (namespace) params.append('namespace', namespace)
-              const ctrl = new AbortController()
-              const tid = setTimeout(() => ctrl.abort(), AGENT_HTTP_TIMEOUT_MS)
-              const res = await fetch(`${LOCAL_AGENT_HTTP_URL}/deployments?${params}`, {
-                signal: ctrl.signal, headers: { Accept: 'application/json' } })
-              clearTimeout(tid)
-              if (!res.ok) return []
-              const data = await res.json().catch(() => null)
-              if (!data) return []
-              return ((data.deployments || []) as Deployment[]).map(d => ({ ...d, cluster: cluster }))
-          })()
-          : await fetchDeploymentsViaAgent(namespace)
-
-        return deriveIssues(deployments)
-      }
-
-      // Fall back to REST API
-      const token = getToken()
-      const hasRealToken = token && token !== 'demo-token'
-      if (hasRealToken && !isBackendUnavailable()) {
-        const data = await fetchAPI<{ issues: DeploymentIssue[] }>('deployment-issues', { cluster, namespace })
-        return data.issues || []
-      }
-
-      throw new Error("No data source available")
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      if (clusterCacheRef.clusters.length > 0 && !isAgentUnavailable()) {
-        const deployments = await fetchDeploymentsViaAgent(namespace, (partialDeps) => {
-          onProgress(deriveIssues(partialDeps))
-        })
-        return deriveIssues(deployments)
-      }
-
-      // Fall back to SSE streaming -> REST per-cluster
-      const issues = await fetchViaSSE<DeploymentIssue>('deployment-issues', 'issues', { namespace }, onProgress)
-      return issues
-    } })
-
-  return {
-    issues: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching deployments with caching
- */
-export function useCachedDeployments(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<Deployment[]> & { deployments: Deployment[] } {
-  const { category = 'deployments' } = options || {}
-  const key = `deployments:${cluster || 'all'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as Deployment[],
-    demoData: getDemoDeployments(),
-    fetcher: async () => {
-      // Try agent first (fast, no backend needed) — skip if agent is unavailable
-      if (clusterCacheRef.clusters.length > 0 && !isAgentUnavailable()) {
-        if (cluster) {
-          const params = new URLSearchParams()
-          const clusterInfo = clusterCacheRef.clusters.find(c => c.name === cluster)
-          params.append('cluster', clusterInfo?.context || cluster)
-          if (namespace) params.append('namespace', namespace)
-
-          const controller = new AbortController()
-          const timeoutId = setTimeout(() => controller.abort(), AGENT_HTTP_TIMEOUT_MS)
-          const response = await fetch(`${LOCAL_AGENT_HTTP_URL}/deployments?${params}`, {
-            signal: controller.signal,
-            headers: { Accept: 'application/json' } })
-          clearTimeout(timeoutId)
-
-          if (response.ok) {
-            const rawData = await response.json().catch(() => null)
-            if (!rawData) return []
-            const data = validateArrayResponse<{ deployments: Deployment[] }>(DeploymentsResponseSchema, rawData, '/agent/deployments', 'deployments')
-            return (data.deployments || []).map(d => ({
-              ...d,
-              cluster: cluster }))
-          }
-        }
-        return fetchDeploymentsViaAgent(namespace)
-      }
-
-      // Fall back to REST API
-      const token = getToken()
-      const hasRealToken = token && token !== 'demo-token'
-      if (hasRealToken && !isBackendUnavailable()) {
-        if (cluster) {
-          const raw = await fetchAPI<unknown>('deployments', { cluster, namespace })
-          const data = validateArrayResponse<{ deployments: Deployment[] }>(DeploymentsResponseSchema, raw, '/api/mcp/deployments', 'deployments')
-          const deployments = data.deployments || []
-          return deployments.map(d => ({ ...d, cluster: d.cluster || cluster }))
-        }
-        return await fetchFromAllClusters<Deployment>('deployments', 'deployments', { namespace })
-      }
-
-      throw new Error("No data source available")
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      if (clusterCacheRef.clusters.length > 0 && !isAgentUnavailable()) {
-        return fetchDeploymentsViaAgent(namespace, onProgress)
-      }
-
-      // Fall back to SSE streaming -> REST per-cluster
-      return await fetchViaSSE<Deployment>('deployments', 'deployments', { namespace }, onProgress)
-    } })
-
-  return {
-    deployments: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching services with caching
- */
-export function useCachedServices(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<Service[]> & { services: Service[] } {
-  const { category = 'services' } = options || {}
-  const key = `services:${cluster || 'all'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as Service[],
-    demoData: getDemoServices(),
-    fetcher: async () => {
-      if (cluster) {
-        const data = await fetchAPI<{ services: Service[] }>('services', { cluster, namespace })
-        return (data.services || []).map(s => ({ ...s, cluster }))
-      }
-      return await fetchFromAllClusters<Service>('services', 'services', { namespace })
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return await fetchViaSSE<Service>('services', 'services', { namespace }, onProgress)
-    } })
-
-  return {
-    services: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
+// Core abort helper (stays here — it's a global singleton operation)
+export { abortAllFetches } from '../lib/cache/fetcherUtils'
+
+// Core workload hooks
+export {
+  useCachedPods,
+  useCachedAllPods,
+  useCachedEvents,
+  useCachedPodIssues,
+  useCachedDeploymentIssues,
+  useCachedDeployments,
+  useCachedServices,
+  useCachedSecurityIssues,
+  useCachedWorkloads,
+} from './useCachedCoreWorkloads'
+
+// Node hooks
+export {
+  useCachedNodes,
+  useCachedAllNodes,
+  useCachedCoreDNSStatus,
+} from './useCachedNodes'
+
+// Node types (re-exported so existing imports still work)
+export type {
+  CoreDNSPodInfo,
+  CoreDNSClusterStatus,
+} from './useCachedNodes'
+
+// GPU hooks
+export {
+  useCachedGPUNodes,
+  useCachedGPUNodeHealth,
+  useGPUHealthCronJob,
+  useCachedHardwareHealth,
+  useCachedWarningEvents,
+} from './useCachedGPU'
+
+// Hardware health types (re-exported so existing imports still work)
+export type {
+  DeviceAlert,
+  DeviceCounts,
+  NodeDeviceInventory,
+  HardwareHealthData,
+} from './useCachedGPU'
+
+// K8s resource hooks
+export {
+  useCachedPVCs,
+  useCachedNamespaces,
+  useCachedJobs,
+  useCachedHPAs,
+  useCachedConfigMaps,
+  useCachedSecrets,
+  useCachedServiceAccounts,
+  useCachedReplicaSets,
+  useCachedStatefulSets,
+  useCachedDaemonSets,
+  useCachedCronJobs,
+  useCachedIngresses,
+  useCachedNetworkPolicies,
+} from './useCachedK8sResources'
+
+// GitOps & RBAC hooks
+export {
+  useCachedHelmReleases,
+  useCachedHelmHistory,
+  useCachedHelmValues,
+  useCachedOperators,
+  useCachedOperatorSubscriptions,
+  useCachedGitOpsDrifts,
+  useCachedBuildpackImages,
+  useCachedK8sRoles,
+  useCachedK8sRoleBindings,
+  useCachedK8sServiceAccounts,
+} from './useCachedGitOps'
 
 // ============================================================================
 // Prow CI Hooks — moved to useCachedProw.ts
@@ -1001,727 +125,48 @@ export * from './useCachedProw'
 
 export * from './useCachedLLMd'
 
-
 // ============================================================================
-// Workloads Cached Hooks
-// ============================================================================
-
-const getDemoWorkloads = (): Workload[] => [
-  { name: 'nginx-ingress', namespace: 'ingress-system', type: 'Deployment', status: 'Running', replicas: 3, readyReplicas: 3, image: 'nginx/nginx-ingress:3.4.0', labels: { app: 'nginx-ingress', tier: 'frontend' }, targetClusters: ['us-east-1', 'us-west-2', 'eu-central-1'], createdAt: new Date(Date.now() - 30 * 86400000).toISOString() },
-  { name: 'api-gateway', namespace: 'production', type: 'Deployment', status: 'Degraded', replicas: 5, readyReplicas: 3, image: 'company/api-gateway:v2.5.1', labels: { app: 'api-gateway', tier: 'api' }, targetClusters: ['us-east-1', 'us-west-2'], createdAt: new Date(Date.now() - 14 * 86400000).toISOString() },
-  { name: 'postgres-primary', namespace: 'databases', type: 'StatefulSet', status: 'Running', replicas: 1, readyReplicas: 1, image: 'postgres:15.4', labels: { app: 'postgres', role: 'primary' }, targetClusters: ['us-east-1'], createdAt: new Date(Date.now() - 60 * 86400000).toISOString() },
-  { name: 'fluentd', namespace: 'logging', type: 'DaemonSet', status: 'Running', replicas: 12, readyReplicas: 12, image: 'fluent/fluentd:v1.16', labels: { app: 'fluentd', tier: 'logging' }, targetClusters: ['us-east-1', 'us-west-2', 'eu-central-1'], createdAt: new Date(Date.now() - 45 * 86400000).toISOString() },
-  { name: 'ml-training', namespace: 'ml-workloads', type: 'Deployment', status: 'Pending', replicas: 1, readyReplicas: 0, image: 'company/ml-trainer:latest', labels: { app: 'ml-training', team: 'data-science' }, targetClusters: ['gpu-cluster-1'], createdAt: new Date(Date.now() - 3600000).toISOString() },
-  { name: 'payment-service', namespace: 'payments', type: 'Deployment', status: 'Failed', replicas: 2, readyReplicas: 0, image: 'company/payment-service:v1.8.0', labels: { app: 'payment-service', tier: 'backend' }, targetClusters: ['us-east-1'], createdAt: new Date(Date.now() - 2 * 86400000).toISOString() },
-]
-
-/** Fetch workloads from the local agent across all clusters */
-async function fetchWorkloadsFromAgent(onProgress?: (partial: Workload[]) => void): Promise<Workload[] | null> {
-  if (isAgentUnavailable()) return null
-
-  const clusters = clusterCacheRef.clusters
-    .filter(c => c.reachable !== false && !c.name.includes('/'))
-  if (clusters.length === 0) return null
-
-  const tasks = clusters.map(({ name, context }) => async () => {
-    const params = new URLSearchParams()
-    params.append('cluster', context || name)
-
-    const ctrl = new AbortController()
-    const tid = setTimeout(() => ctrl.abort(), AGENT_HTTP_TIMEOUT_MS)
-    const res = await fetch(`${LOCAL_AGENT_HTTP_URL}/deployments?${params}`, {
-      signal: ctrl.signal,
-      headers: { Accept: 'application/json' } })
-    clearTimeout(tid)
-
-    if (!res.ok) throw new Error(`Agent ${res.status}`)
-    const data = await res.json().catch(() => null)
-    if (!data) throw new Error('Invalid JSON response from agent')
-    return ((data.deployments || []) as Array<Record<string, unknown>>).map(d => {
-      const st = String(d.status || 'running')
-      let ws: Workload['status'] = 'Running'
-      if (st === 'failed') ws = 'Failed'
-      else if (st === 'deploying') ws = 'Pending'
-      else if (Number(d.readyReplicas || 0) < Number(d.replicas || 1)) ws = 'Degraded'
-      return {
-        name: String(d.name || ''),
-        namespace: String(d.namespace || 'default'),
-        type: 'Deployment' as const,
-        cluster: name,
-        targetClusters: [name],
-        replicas: Number(d.replicas || 1),
-        readyReplicas: Number(d.readyReplicas || 0),
-        status: ws,
-        image: String(d.image || ''),
-        createdAt: new Date().toISOString() }
-    })
-  })
-
-  const accumulated: Workload[] = []
-  function handleSettled(result: PromiseSettledResult<Workload[]>) {
-    if (result.status === 'fulfilled') {
-      accumulated.push(...result.value)
-      onProgress?.([...accumulated])
-    }
-  }
-  await settledWithConcurrency(tasks, undefined, handleSettled)
-  return accumulated.length > 0 ? accumulated : null
-}
-
-/**
- * Hook for fetching workloads with caching.
- * Fetches all workloads across all clusters via agent, then REST fallback.
- */
-export function useCachedWorkloads(
-  options?: { category?: RefreshCategory }
-): CachedHookResult<Workload[]> & { workloads: Workload[] } {
-  const { category = 'deployments' } = options || {}
-  const key = 'workloads:all:all'
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as Workload[],
-    demoData: getDemoWorkloads(),
-    fetcher: async () => {
-      // Try agent first (fast, no backend needed)
-      const agentData = await fetchWorkloadsFromAgent()
-      if (agentData) return agentData
-
-      // Fall back to REST API
-      const token = getToken()
-      const hasRealToken = token && token !== 'demo-token'
-      if (hasRealToken && !isBackendUnavailable()) {
-        const res = await fetch('/api/workloads', {
-          headers: {
-            'Content-Type': 'application/json',
-            Authorization: `Bearer ${token}` },
-          signal: AbortSignal.timeout(FETCH_DEFAULT_TIMEOUT_MS) })
-        if (res.ok) {
-          const data = await res.json().catch(() => null)
-          if (!data) return []
-          const items = (data.items || data) as Array<Record<string, unknown>>
-          return items.map(d => ({
-            name: String(d.name || ''),
-            namespace: String(d.namespace || 'default'),
-            type: (String(d.type || 'Deployment')) as Workload['type'],
-            cluster: String(d.cluster || ''),
-            targetClusters: (d.targetClusters as string[]) || (d.cluster ? [String(d.cluster)] : []),
-            replicas: Number(d.replicas || 1),
-            readyReplicas: Number(d.readyReplicas || 0),
-            status: (String(d.status || 'Running')) as Workload['status'],
-            image: String(d.image || ''),
-            labels: (d.labels as Record<string, string>) || {},
-            createdAt: String(d.createdAt || new Date().toISOString()) }))
-        }
-      }
-
-      return []
-    },
-    progressiveFetcher: async (onProgress) => {
-      // Try agent first (progressive via kc-agent)
-      const agentData = await fetchWorkloadsFromAgent(onProgress)
-      if (agentData) return agentData
-
-      // Fall back to SSE streaming -> progressive per-cluster
-      return await fetchViaSSE<Workload>('workloads', 'workloads', {}, onProgress)
-    } })
-
-  return {
-    workloads: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-// ============================================================================
-// Security Cached Hooks
+// ISO 27001 Security Audit — re-exported from useCachedISO27001.ts
 // ============================================================================
 
-/**
- * Fetch security issues via kubectlProxy - scans pods for security misconfigurations
- */
-async function fetchSecurityIssuesViaKubectl(cluster?: string, namespace?: string, onProgress?: (partial: SecurityIssue[]) => void): Promise<SecurityIssue[]> {
-  if (isAgentUnavailable()) return []
-  const clusters = getAgentClusters()
-  if (clusters.length === 0) return []
-
-  const severityOrder: Record<string, number> = { critical: 0, high: 1, medium: 2, low: 3, info: 4 }
-
-  const tasks = clusters
-    .filter(c => !cluster || c.name === cluster)
-    .map(({ name, context }) => async () => {
-      const ctx = context || name
-      // Get all pods and check for security issues
-      const nsFlag = namespace ? ['-n', namespace] : ['-A']
-      const response = await kubectlProxy.exec(
-        ['get', 'pods', ...nsFlag, '-o', 'json'],
-        { context: ctx, timeout: KUBECTL_EXTENDED_TIMEOUT_MS }
-      )
-
-      if (response.exitCode !== 0) return []
-
-      const data = JSON.parse(response.output)
-      const issues: SecurityIssue[] = []
-
-      for (const pod of data.items || []) {
-        const podName = pod.metadata?.name || 'unknown'
-        const podNs = pod.metadata?.namespace || 'default'
-        const spec = pod.spec || {}
-
-        // Check for security misconfigurations
-        for (const container of spec.containers || []) {
-          const sc = container.securityContext || {}
-          const podSc = spec.securityContext || {}
-
-          // Privileged container
-          if (sc.privileged === true) {
-            issues.push({ name: podName, namespace: podNs, cluster: name, issue: 'Privileged container', severity: 'high', details: 'Container running in privileged mode' })
-          }
-
-          // Running as root
-          if (sc.runAsUser === 0 || (sc.runAsNonRoot !== true && podSc.runAsNonRoot !== true && !sc.runAsUser)) {
-            const isRoot = sc.runAsUser === 0 || podSc.runAsUser === 0
-            if (isRoot) {
-              issues.push({ name: podName, namespace: podNs, cluster: name, issue: 'Running as root', severity: 'high', details: 'Container running as root user' })
-            }
-          }
-
-          // Missing security context
-          if (!sc.runAsNonRoot && !sc.readOnlyRootFilesystem && !sc.allowPrivilegeEscalation && !sc.capabilities) {
-            issues.push({ name: podName, namespace: podNs, cluster: name, issue: 'Missing security context', severity: 'low', details: 'No security context defined' })
-          }
-
-          // Capabilities not dropped
-          if (sc.capabilities?.drop?.length === 0 || !sc.capabilities?.drop) {
-            if (sc.capabilities?.add?.length > 0) {
-              issues.push({ name: podName, namespace: podNs, cluster: name, issue: 'Capabilities not dropped', severity: 'medium', details: 'Container not dropping all capabilities' })
-            }
-          }
-        }
-
-        // Host network
-        if (spec.hostNetwork === true) {
-          issues.push({ name: podName, namespace: podNs, cluster: name, issue: 'Host network enabled', severity: 'medium', details: 'Pod using host network namespace' })
-        }
-
-        // Host PID
-        if (spec.hostPID === true) {
-          issues.push({ name: podName, namespace: podNs, cluster: name, issue: 'Host PID enabled', severity: 'high', details: 'Pod using host PID namespace' })
-        }
-
-        // Host IPC
-        if (spec.hostIPC === true) {
-          issues.push({ name: podName, namespace: podNs, cluster: name, issue: 'Host IPC enabled', severity: 'medium', details: 'Pod using host IPC namespace' })
-        }
-      }
-
-      return issues
-    })
-
-  const accumulated: SecurityIssue[] = []
-  function handleSettled(result: PromiseSettledResult<SecurityIssue[]>) {
-    if (result.status === 'fulfilled') {
-      accumulated.push(...result.value)
-      accumulated.sort((a, b) => (severityOrder[a.severity] || 5) - (severityOrder[b.severity] || 5))
-      onProgress?.([...accumulated])
-    }
-  }
-  await settledWithConcurrency(tasks, undefined, handleSettled)
-  // Final sort
-  return accumulated.sort((a, b) => (severityOrder[a.severity] || 5) - (severityOrder[b.severity] || 5))
-}
-
-/**
- * Hook for fetching security issues with caching
- * Provides stale-while-revalidate: shows cached data immediately while refreshing
- */
-export function useCachedSecurityIssues(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<SecurityIssue[]> & { issues: SecurityIssue[] } {
-  const { category = 'pods' } = options || {}
-  const key = `securityIssues:${cluster || 'all'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as SecurityIssue[],
-    demoData: getDemoSecurityIssues(),
-    fetcher: async () => {
-      // Try kubectl proxy first (uses agent to run kubectl commands) — skip if agent is unavailable
-      if (clusterCacheRef.clusters.length > 0 && !isAgentUnavailable()) {
-        try {
-          const issues = await fetchSecurityIssuesViaKubectl(cluster, namespace)
-          if (issues.length > 0) return issues
-        } catch (err) {
-          console.error('[useCachedSecurityIssues] kubectl fetch failed:', err)
-        }
-      }
-
-      // Fall back to REST API
-      const token = getToken()
-      const hasRealToken = token && token !== 'demo-token'
-      if (hasRealToken && !isBackendUnavailable()) {
-        try {
-          const params = new URLSearchParams()
-          if (cluster) params.append('cluster', cluster)
-          if (namespace) params.append('namespace', namespace)
-          const response = await authFetch(`${LOCAL_AGENT_HTTP_URL}/security-issues?${params}`, {
-            method: 'GET',
-            headers: {
-              'Content-Type': 'application/json',
-              Authorization: `Bearer ${token}` },
-            signal: AbortSignal.timeout(FETCH_DEFAULT_TIMEOUT_MS) })
-          if (response.ok) {
-            const rawSecurity = await response.json().catch(() => null)
-            const data = validateResponse(SecurityIssuesResponseSchema, rawSecurity, '/security-issues')
-            if (data?.issues && data.issues.length > 0) return data.issues
-          }
-        } catch (err) {
-          console.error('[useCachedSecurityIssues] API fetch failed:', err)
-        }
-      }
-
-      throw new Error("No data source available")
-    },
-    // Progressive loading: show results as each cluster completes
-    progressiveFetcher: !cluster ? async (onProgress) => {
-      // Try kubectl proxy first (progressive) — skip if agent is unavailable
-      if (clusterCacheRef.clusters.length > 0 && !isAgentUnavailable()) {
-        try {
-          const issues = await fetchSecurityIssuesViaKubectl(cluster, namespace, onProgress)
-          if (issues.length > 0) return issues
-        } catch (err) {
-          console.error('[useCachedSecurityIssues] progressive kubectl fetch failed:', err)
-        }
-      }
-
-      // Fall back to SSE streaming -> REST per-cluster
-      return await fetchViaSSE<SecurityIssue>('security-issues', 'issues', { namespace }, onProgress)
-    } : undefined })
-
-  return {
-    issues: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-// ============================================================================
-// Nodes Cached Hooks (SSE-enabled)
-// ============================================================================
-
-const getDemoCachedNodes = (): NodeInfo[] => [
-  { name: 'node-1', cluster: 'prod-east', status: 'Ready', roles: ['control-plane', 'master'], kubeletVersion: 'v1.28.4', cpuCapacity: '8', memoryCapacity: '32Gi', podCapacity: '110', conditions: [{ type: 'Ready', status: 'True' }], unschedulable: false },
-  { name: 'node-2', cluster: 'prod-east', status: 'Ready', roles: ['worker'], kubeletVersion: 'v1.28.4', cpuCapacity: '16', memoryCapacity: '64Gi', podCapacity: '110', conditions: [{ type: 'Ready', status: 'True' }], unschedulable: false },
-  { name: 'gpu-node-1', cluster: 'vllm-d', status: 'Ready', roles: ['worker'], kubeletVersion: 'v1.28.4', cpuCapacity: '32', memoryCapacity: '128Gi', podCapacity: '110', conditions: [{ type: 'Ready', status: 'True' }], unschedulable: false },
-  { name: 'kind-control-plane', cluster: 'kind-local', status: 'Ready', roles: ['control-plane'], kubeletVersion: 'v1.27.3', cpuCapacity: '4', memoryCapacity: '8Gi', podCapacity: '110', conditions: [{ type: 'Ready', status: 'True' }], unschedulable: false },
-]
-
-/**
- * Hook for fetching nodes with caching and SSE streaming.
- * When no cluster is specified, fetches from all available clusters via SSE.
- */
-export function useCachedNodes(
-  cluster?: string,
-): CachedHookResult<NodeInfo[]> & { nodes: NodeInfo[] } {
-  const key = `nodes:${cluster || 'all'}`
-
-  const result = useCache({
-    key,
-    category: 'pods' as RefreshCategory,
-    initialData: [] as NodeInfo[],
-    demoData: getDemoCachedNodes(),
-    demoWhenEmpty: !cluster,
-    persist: true,
-    fetcher: async () => {
-      if (cluster) {
-        const raw = await fetchAPI<unknown>('nodes', { cluster })
-        const data = validateArrayResponse<{ nodes: NodeInfo[] }>(NodesResponseSchema, raw, '/api/mcp/nodes', 'nodes')
-        return (data.nodes || []).map(n => ({ ...n, cluster }))
-      }
-      return fetchFromAllClusters<NodeInfo>('nodes', 'nodes', {})
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return fetchViaSSE<NodeInfo>('nodes', 'nodes', {}, onProgress)
-    } })
-
-  // Only report demo fallback when NOT loading — prevents Demo badge
-  // from flashing during the optimistic demo phase while a real fetch
-  // is still in-flight (CLAUDE.md: effectiveIsDemoFallback pattern).
-  const effectiveIsDemoFallback = result.isDemoFallback && !result.isLoading
-
-  return {
-    nodes: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: effectiveIsDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching a cumulative, cross-cluster node list for the landing
- * dashboard's "Nodes" stat-block drill-down (Issue 8840).
- *
- * Differs from {@link useCachedNodes} (no cluster arg) in two ways:
- *
- * 1. It iterates DEDUPLICATED clusters from the shared cluster cache via
- *    {@link deduplicateClustersByServer}. The generic fan-out in
- *    {@link fetchFromAllClusters} only filters out long-context aliases by
- *    name, which can still double-count when two short-named contexts point
- *    to the same API server. See MEMORY.md: "ALWAYS use
- *    DeduplicatedClusters() when iterating clusters".
- *
- * 2. It DOES NOT fall back to demo data on empty live results. The drill-down
- *    is expected to render an authoritative cross-cluster list (or the
- *    existing "summary reports N nodes but list is empty" explainer),
- *    NOT four hard-coded fake nodes. The parent stat block (on the Dashboard)
- *    retains the original {@link useCachedNodes} optimistic-demo behaviour so
- *    tiles never flash empty. Only the drill-down is switched to this hook.
- *
- * An `isDemoFallback` field is still returned (always `false` in live mode,
- * `true` only when the whole app is in demo mode) so the drill-down can keep
- * its existing `nodesIsDemoFallback` wiring intact — see the task's
- * `isDemoData` preservation rule.
- */
-export function useCachedAllNodes(): CachedHookResult<NodeInfo[]> & { nodes: NodeInfo[] } {
-  const key = 'nodes:all:dedup'
-
-  const result = useCache({
-    key,
-    category: 'pods' as RefreshCategory,
-    initialData: [] as NodeInfo[],
-    demoData: getDemoCachedNodes(),
-    // IMPORTANT: no demoWhenEmpty — we never want to mask a real empty /
-    // partially-failed cross-cluster result with four hard-coded fake nodes
-    // in the drill-down list. The drill-down has its own empty-state
-    // explainer (RBAC hint, summary-vs-list mismatch) that is much more
-    // useful than synthetic demo data.
-    persist: true,
-    fetcher: async () => {
-      const allClusters = clusterCacheRef.clusters || []
-      // Apply the same deduplication that useClusters().deduplicatedClusters
-      // uses — multiple kubeconfig contexts can point to the same API
-      // server, and without dedup each node would be listed once per
-      // context. See MEMORY.md: feedback_dedupe_clusters.md.
-      const deduped = deduplicateClustersByServer(allClusters)
-      const reachable = deduped.filter(
-        (c) => c.reachable !== false && !c.name.includes('/'),
-      )
-      if (reachable.length === 0) {
-        throw new Error(
-          'No reachable clusters (agent connecting or backend not authenticated)',
-        )
-      }
-
-      // Fan out per-cluster fetches in parallel. Each per-cluster call is
-      // identical to what useCachedNodes(clusterName) would do — using the
-      // same /api/mcp/nodes endpoint that already works on the per-cluster
-      // drill-down, so the "live data path exists" invariant from the
-      // issue description holds.
-      const tasks = reachable.map((cluster) => async () => {
-        try {
-          const raw = await fetchAPI<unknown>('nodes', { cluster: cluster.name })
-          const data = validateArrayResponse<{ nodes: NodeInfo[] }>(
-            NodesResponseSchema,
-            raw,
-            '/api/mcp/nodes',
-            'nodes',
-          )
-          return (data.nodes || []).map((n) => ({ ...n, cluster: cluster.name }))
-        } catch {
-          // Per-cluster failure: tolerate it so a single unreachable cluster
-          // doesn't wipe the whole aggregate. Accumulated nodes from other
-          // clusters still render; the drill-down's empty-state explainer
-          // handles the all-failed case.
-          return [] as NodeInfo[]
-        }
-      })
-
-      const accumulated: NodeInfo[] = []
-      async function handleSettled(res: PromiseSettledResult<NodeInfo[]>) {
-        if (res.status === 'fulfilled') {
-          accumulated.push(...res.value)
-        }
-      }
-      await settledWithConcurrency(tasks, undefined, handleSettled)
-      return accumulated
-    } })
-
-  return {
-    nodes: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-// ============================================================================
-// GPU Node Health Cached Hooks (SSE-enabled, proactive monitoring)
-// ============================================================================
-
-const getDemoCachedGPUNodeHealth = (): GPUNodeHealthStatus[] => [
-  {
-    nodeName: 'gpu-node-1', cluster: 'vllm-gpu-cluster', status: 'healthy',
-    gpuCount: 8, gpuType: 'NVIDIA A100-SXM4-80GB',
-    checks: [
-      { name: 'node_ready', passed: true },
-      { name: 'scheduling', passed: true },
-      { name: 'gpu-feature-discovery', passed: true },
-      { name: 'nvidia-device-plugin', passed: true },
-      { name: 'dcgm-exporter', passed: true },
-      { name: 'stuck_pods', passed: true },
-      { name: 'gpu_events', passed: true },
-    ],
-    issues: [], stuckPods: 0, checkedAt: new Date().toISOString() },
-  {
-    nodeName: 'gpu-node-2', cluster: 'vllm-gpu-cluster', status: 'degraded',
-    gpuCount: 8, gpuType: 'NVIDIA A100-SXM4-80GB',
-    checks: [
-      { name: 'node_ready', passed: true },
-      { name: 'scheduling', passed: true },
-      { name: 'gpu-feature-discovery', passed: false, message: 'CrashLoopBackOff (12 restarts)' },
-      { name: 'nvidia-device-plugin', passed: true },
-      { name: 'dcgm-exporter', passed: true },
-      { name: 'stuck_pods', passed: true },
-      { name: 'gpu_events', passed: true },
-    ],
-    issues: ['gpu-feature-discovery: CrashLoopBackOff (12 restarts)'], stuckPods: 0, checkedAt: new Date().toISOString() },
-  {
-    nodeName: 'gpu-node-3', cluster: 'eks-prod-us-east-1', status: 'unhealthy',
-    gpuCount: 4, gpuType: 'NVIDIA V100',
-    checks: [
-      { name: 'node_ready', passed: false, message: 'Node is NotReady' },
-      { name: 'scheduling', passed: false, message: 'Node is cordoned (SchedulingDisabled)' },
-      { name: 'gpu-feature-discovery', passed: false, message: 'CrashLoopBackOff (128 restarts)' },
-      { name: 'nvidia-device-plugin', passed: false, message: 'CrashLoopBackOff (64 restarts)' },
-      { name: 'dcgm-exporter', passed: true },
-      { name: 'stuck_pods', passed: false, message: '54 pods stuck (ContainerStatusUnknown/Terminating)' },
-      { name: 'gpu_events', passed: false, message: '3 GPU warning events in last hour' },
-    ],
-    issues: ['Node is NotReady', 'Node is cordoned', 'gpu-feature-discovery: CrashLoopBackOff (128 restarts)', '54 pods stuck'],
-    stuckPods: 54, checkedAt: new Date().toISOString() },
-]
-
-/**
- * Hook for fetching GPU node health data with caching and SSE streaming.
- */
-export function useCachedGPUNodeHealth(
-  cluster?: string,
-): CachedHookResult<GPUNodeHealthStatus[]> & { nodes: GPUNodeHealthStatus[] } {
-  const key = `gpu-node-health:${cluster || 'all'}`
-
-  const result = useCache({
-    key,
-    category: 'pods' as RefreshCategory,
-    initialData: [] as GPUNodeHealthStatus[],
-    demoData: getDemoCachedGPUNodeHealth(),
-    persist: true,
-    fetcher: async () => {
-      if (cluster) {
-        const raw = await fetchAPI<unknown>('gpu-nodes/health', { cluster })
-        const data = validateArrayResponse<{ nodes: GPUNodeHealthStatus[] }>(GPUNodeHealthResponseSchema, raw, '/api/mcp/gpu-nodes/health', 'nodes')
-        return (data.nodes || []).map(n => ({ ...n, cluster }))
-      }
-      return fetchFromAllClusters<GPUNodeHealthStatus>('gpu-nodes/health', 'nodes', {})
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return fetchViaSSE<GPUNodeHealthStatus>('gpu-nodes/health', 'nodes', {}, onProgress)
-    } })
-
-  return {
-    nodes: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-// ============================================================================
-// GPU Health CronJob Management Hook
-// ============================================================================
-
-/**
- * Hook for managing GPU health CronJob installation per cluster.
- * Uses useCache for persistent status caching; install/uninstall are imperative actions.
- */
-export function useGPUHealthCronJob(cluster?: string) {
-  const key = `gpu-cronjob-status:${cluster || 'none'}`
-  const [actionInProgress, setActionInProgress] = useState<'install' | 'uninstall' | null>(null)
-  const [actionError, setActionError] = useState<string | null>(null)
-
-  const result = useCache<GPUHealthCronJobStatus | null>({
-    key,
-    category: 'pods' as RefreshCategory,
-    initialData: null,
-    demoData: null,
-    persist: true,
-    enabled: !!cluster,
-    fetcher: async () => {
-      if (!cluster) return null
-      return fetchAPI<GPUHealthCronJobStatus>('gpu-nodes/health/cronjob', { cluster })
-    } })
-
-  const install = async (opts?: { namespace?: string; schedule?: string; tier?: number }) => {
-    if (!cluster) return
-    setActionInProgress('install')
-    setActionError(null)
-    try {
-      const token = getToken()
-      if (!token) throw new Error('No authentication token')
-      // #7993 Phase 3e: GPU health cronjob install is a user-initiated
-      // tooling install. It runs through kc-agent under the user's own
-      // kubeconfig instead of the backend pod ServiceAccount.
-      const response = await fetch(`${LOCAL_AGENT_HTTP_URL}/gpu-health-cronjob`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${token}` },
-        body: JSON.stringify({
-          cluster,
-          namespace: opts?.namespace,
-          schedule: opts?.schedule,
-          tier: opts?.tier ?? 2 }),
-        signal: AbortSignal.timeout(FETCH_DEFAULT_TIMEOUT_MS) })
-      if (!response.ok) {
-        const text = await response.text()
-        throw new Error(text || `Install failed: ${response.status}`)
-      }
-      await result.refetch()
-    } catch (err) {
-      setActionError(err instanceof Error ? err.message : 'Failed to install CronJob')
-    } finally {
-      setActionInProgress(null)
-    }
-  }
-
-  const uninstall = async (opts?: { namespace?: string }) => {
-    if (!cluster) return
-    setActionInProgress('uninstall')
-    setActionError(null)
-    try {
-      const token = getToken()
-      if (!token) throw new Error('No authentication token')
-      // #7993 Phase 3e: GPU health cronjob uninstall goes through kc-agent
-      // under the user's own kubeconfig.
-      const response = await fetch(`${LOCAL_AGENT_HTTP_URL}/gpu-health-cronjob`, {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${token}` },
-        body: JSON.stringify({
-          cluster,
-          namespace: opts?.namespace }),
-        signal: AbortSignal.timeout(FETCH_DEFAULT_TIMEOUT_MS) })
-      if (!response.ok) {
-        const text = await response.text()
-        throw new Error(text || `Uninstall failed: ${response.status}`)
-      }
-      await result.refetch()
-    } catch (err) {
-      setActionError(err instanceof Error ? err.message : 'Failed to uninstall CronJob')
-    } finally {
-      setActionInProgress(null)
-    }
-  }
-
-  return {
-    status: result.data,
-    isLoading: result.isLoading,
-    error: actionError || result.error,
-    actionInProgress,
-    install,
-    uninstall,
-    refetch: result.refetch }
-}
-
-// ============================================================================
-// Warning Events Cached Hooks (SSE-enabled)
-// ============================================================================
-
-const getDemoCachedWarningEvents = (): ClusterEvent[] => [
-  { type: 'Warning', reason: 'FailedScheduling', message: 'Insufficient cpu', namespace: 'production', object: 'Pod/api-gateway-7d9c8b7f5-abcde', count: 3, firstSeen: new Date(Date.now() - 300000).toISOString(), lastSeen: new Date().toISOString(), cluster: 'prod-east' },
-  { type: 'Warning', reason: 'BackOff', message: 'Back-off restarting failed container', namespace: 'monitoring', object: 'Pod/prometheus-agent-0', count: 5, firstSeen: new Date(Date.now() - 600000).toISOString(), lastSeen: new Date().toISOString(), cluster: 'prod-east' },
-  { type: 'Warning', reason: 'FailedCreate', message: 'Error creating: pods "worker-xyz" is forbidden', namespace: 'ml-workloads', object: 'Job/training-job-123', count: 1, firstSeen: new Date(Date.now() - 120000).toISOString(), lastSeen: new Date().toISOString(), cluster: 'vllm-d' },
-]
-
-/**
- * Hook for fetching warning events with caching and SSE streaming.
- * When no cluster is specified, fetches from all clusters via SSE.
- */
-export function useCachedWarningEvents(
-  cluster?: string,
-  namespace?: string,
-  options?: { limit?: number; category?: RefreshCategory }
-): CachedHookResult<ClusterEvent[]> & { events: ClusterEvent[] } {
-  const { limit = 50, category = 'realtime' } = options || {}
-  const key = `warningEvents:${cluster || 'all'}:${namespace || 'all'}:${limit}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as ClusterEvent[],
-    demoData: getDemoCachedWarningEvents(),
-    fetcher: async () => {
-      if (cluster) {
-        const data = await fetchAPI<{ events: ClusterEvent[] }>('events/warnings', { cluster, namespace, limit })
-        return (data.events || []).map(e => ({ ...e, cluster }))
-      }
-      const events = await fetchFromAllClusters<ClusterEvent>('events/warnings', 'events', { namespace, limit })
-      return events.slice(0, limit)
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      const events = await fetchViaSSE<ClusterEvent>('events/warnings', 'events', { namespace, limit }, (partial) => {
-        onProgress(partial.slice(0, limit))
-      })
-      return events.slice(0, limit)
-    } })
-
-  return {
-    events: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
+export * from './useCachedISO27001'
 
 // ============================================================================
 // Standalone fetchers for prefetch (no React hooks, plain async)
 // ============================================================================
+
+import { isBackendUnavailable, authFetch } from '../lib/api'
+import { clusterCacheRef } from './mcp/shared'
+import { isAgentUnavailable } from './useLocalAgent'
+import { LOCAL_AGENT_HTTP_URL } from '../lib/constants'
+import { FETCH_DEFAULT_TIMEOUT_MS } from '../lib/constants/network'
+import {
+  fetchAPI,
+  fetchFromAllClusters,
+  getToken,
+  MAX_PREFETCH_PODS,
+} from '../lib/cache/fetcherUtils'
+import {
+  fetchPodIssuesViaAgent,
+  fetchDeploymentsViaAgent,
+  fetchWorkloadsFromAgent,
+  fetchSecurityIssuesViaKubectl,
+} from './useCachedCoreWorkloads'
+import { fetchProwJobs } from './useCachedProw'
+import { fetchLLMdServers, fetchLLMdModels } from './useCachedLLMd'
+import { validateResponse } from '../lib/schemas/validate'
+import { SecurityIssuesResponseSchema } from '../lib/schemas'
+import type {
+  PodInfo,
+  PodIssue,
+  ClusterEvent,
+  DeploymentIssue,
+  Deployment,
+  Service,
+  SecurityIssue,
+  NodeInfo,
+} from './useMCP'
+import type { Workload } from './useWorkloads'
 
 /** Core data fetchers — used by prefetchCardData to warm caches at startup */
 export const coreFetchers = {
@@ -1839,1428 +284,8 @@ export const coreFetchers = {
     return []
   } }
 
-// ============================================================================
-// Hardware Health (device alerts + inventory)
-// ============================================================================
-
-// Device alert from agent
-export interface DeviceAlert {
-  id: string
-  nodeName: string
-  cluster: string
-  deviceType: string
-  previousCount: number
-  currentCount: number
-  droppedCount: number
-  firstSeen: string
-  lastSeen: string
-  severity: string
-}
-
-interface DeviceAlertsResponse {
-  alerts: DeviceAlert[]
-  nodeCount: number
-  timestamp: string
-}
-
-export interface DeviceCounts {
-  gpuCount: number
-  nicCount: number
-  nvmeCount: number
-  infinibandCount: number
-  sriovCapable: boolean
-  rdmaAvailable: boolean
-  mellanoxPresent: boolean
-  nvidiaNicPresent: boolean
-  spectrumScale: boolean
-  mofedReady: boolean
-  gpuDriverReady: boolean
-}
-
-export interface NodeDeviceInventory {
-  nodeName: string
-  cluster: string
-  devices: DeviceCounts
-  lastSeen: string
-}
-
-interface DeviceInventoryResponse {
-  nodes: NodeDeviceInventory[]
-  timestamp: string
-}
-
-export interface HardwareHealthData {
-  alerts: DeviceAlert[]
-  inventory: NodeDeviceInventory[]
-  nodeCount: number
-  lastUpdate: string | null
-}
-
-const DEMO_HW_ALERTS: DeviceAlert[] = [
-  {
-    id: 'demo-1',
-    nodeName: 'gpu-node-1',
-    cluster: 'production',
-    deviceType: 'gpu',
-    previousCount: 8,
-    currentCount: 6,
-    droppedCount: 2,
-    firstSeen: new Date().toISOString(),
-    lastSeen: new Date().toISOString(),
-    severity: 'critical' },
-  {
-    id: 'demo-2',
-    nodeName: 'gpu-node-2',
-    cluster: 'production',
-    deviceType: 'infiniband',
-    previousCount: 2,
-    currentCount: 1,
-    droppedCount: 1,
-    firstSeen: new Date().toISOString(),
-    lastSeen: new Date().toISOString(),
-    severity: 'warning' },
-]
-
-const DEMO_HW_INVENTORY: NodeDeviceInventory[] = [
-  {
-    nodeName: 'gpu-node-1',
-    cluster: 'production',
-    devices: { gpuCount: 8, nicCount: 2, nvmeCount: 4, infinibandCount: 2, sriovCapable: true, rdmaAvailable: true, mellanoxPresent: true, nvidiaNicPresent: false, spectrumScale: false, mofedReady: true, gpuDriverReady: true },
-    lastSeen: new Date().toISOString() },
-  {
-    nodeName: 'gpu-node-2',
-    cluster: 'production',
-    devices: { gpuCount: 8, nicCount: 2, nvmeCount: 4, infinibandCount: 2, sriovCapable: true, rdmaAvailable: true, mellanoxPresent: true, nvidiaNicPresent: false, spectrumScale: false, mofedReady: true, gpuDriverReady: true },
-    lastSeen: new Date().toISOString() },
-  {
-    nodeName: 'compute-node-1',
-    cluster: 'staging',
-    devices: { gpuCount: 0, nicCount: 1, nvmeCount: 2, infinibandCount: 0, sriovCapable: false, rdmaAvailable: false, mellanoxPresent: false, nvidiaNicPresent: false, spectrumScale: false, mofedReady: false, gpuDriverReady: false },
-    lastSeen: new Date().toISOString() },
-]
-
-const HW_INITIAL_DATA: HardwareHealthData = {
-  alerts: [],
-  inventory: [],
-  nodeCount: 0,
-  lastUpdate: null }
-
-const HW_DEMO_DATA: HardwareHealthData = {
-  alerts: DEMO_HW_ALERTS,
-  inventory: DEMO_HW_INVENTORY,
-  nodeCount: DEMO_HW_INVENTORY.length,
-  lastUpdate: new Date().toISOString() }
-
-async function fetchHardwareHealth(): Promise<HardwareHealthData> {
-  const controller = new AbortController()
-  const timeoutId = setTimeout(() => controller.abort(), AI_PREDICTION_TIMEOUT_MS)
-
-  try {
-    const [alertsRes, inventoryRes] = await Promise.all([
-      fetch(`${LOCAL_AGENT_HTTP_URL}/devices/alerts`, {
-        method: 'GET',
-        headers: { 'Accept': 'application/json' },
-        signal: controller.signal }).catch(() => null),
-      fetch(`${LOCAL_AGENT_HTTP_URL}/devices/inventory`, {
-        method: 'GET',
-        headers: { 'Accept': 'application/json' },
-        signal: controller.signal }).catch(() => null),
-    ])
-    clearTimeout(timeoutId)
-
-    const result: HardwareHealthData = {
-      alerts: [],
-      inventory: [],
-      nodeCount: 0,
-      lastUpdate: new Date().toISOString() }
-
-    if (alertsRes?.ok) {
-      const data = await alertsRes.json().catch(() => null) as DeviceAlertsResponse | null
-      if (data) {
-        result.alerts = data.alerts || []
-        result.nodeCount = data.nodeCount
-      }
-    }
-
-    if (inventoryRes?.ok) {
-      const data = await inventoryRes.json().catch(() => null) as DeviceInventoryResponse | null
-      if (data) {
-        result.inventory = data.nodes || []
-        if (data.nodes && data.nodes.length > 0) {
-          result.nodeCount = data.nodes.length
-        }
-      }
-    }
-
-    // If neither endpoint returned data, throw so useCache tracks the failure
-    if (!alertsRes?.ok && !inventoryRes?.ok) {
-      throw new Error('Device endpoints unavailable')
-    }
-
-    return result
-  } catch (e) {
-    clearTimeout(timeoutId)
-    throw e
-  }
-}
-
-/**
- * Hook for fetching hardware health data (device alerts + inventory) with caching.
- * Uses IndexedDB persistence so data survives navigation.
- */
-export function useCachedHardwareHealth(): CachedHookResult<HardwareHealthData> {
-  const result = useCache({
-    key: 'hardware-health',
-    category: 'pods', // 30-second refresh
-    initialData: HW_INITIAL_DATA,
-    demoData: HW_DEMO_DATA,
-    persist: true,
-    // Don't gate on isAgentUnavailable() — the agent may connect after the hook
-    // mounts and `enabled` is only read once. The fetcher handles unavailability
-    // internally by throwing, which useCache tracks as consecutive failures.
-    fetcher: fetchHardwareHealth })
-
-  return {
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
 /** Specialty data fetchers — lower priority, prefetched after core data */
 export const specialtyFetchers = {
   prowJobs: () => fetchProwJobs('prow', 'prow'),
   llmdServers: () => fetchLLMdServers(['vllm-d', 'platform-eval']),
   llmdModels: () => fetchLLMdModels(['vllm-d', 'platform-eval']) }
-
-// -- CoreDNS status --
-
-export interface CoreDNSPodInfo {
-  name: string
-  status: string
-  ready: string
-  restarts: number
-  version: string
-}
-
-export interface CoreDNSClusterStatus {
-  cluster: string
-  pods: CoreDNSPodInfo[]
-  healthy: boolean
-  totalRestarts: number
-}
-
-const getDemoCoreDNSStatus = (): CoreDNSClusterStatus[] => [
-  {
-    cluster: 'eks-prod-us-east-1',
-    pods: [
-      { name: 'coredns-7db6d8ff4d-xk2p8', status: 'Running', ready: '1/1', restarts: 0, version: '1.11.1' },
-      { name: 'coredns-7db6d8ff4d-n9wq3', status: 'Running', ready: '1/1', restarts: 0, version: '1.11.1' },
-    ],
-    healthy: true,
-    totalRestarts: 0 },
-  {
-    cluster: 'gke-staging',
-    pods: [
-      { name: 'coredns-6d4b75cb6d-abcde', status: 'Running', ready: '1/1', restarts: 2, version: '1.10.1' },
-      { name: 'coredns-6d4b75cb6d-fghij', status: 'Running', ready: '1/1', restarts: 0, version: '1.10.1' },
-    ],
-    healthy: true,
-    totalRestarts: 2 },
-  {
-    cluster: 'aks-dev-westeu',
-    pods: [
-      { name: 'coredns-abc123-xyz99', status: 'CrashLoopBackOff', ready: '0/1', restarts: 7, version: '1.9.3' },
-    ],
-    healthy: false,
-    totalRestarts: 7 },
-]
-
-// fetches coredns pods from kube-system and builds per-cluster health info
-export function useCachedCoreDNSStatus(
-  cluster?: string
-): CachedHookResult<CoreDNSClusterStatus[]> & { clusters: CoreDNSClusterStatus[] } {
-  const key = `coredns:${cluster || 'all'}`
-
-  const result = useCache({
-    key,
-    category: 'pods' as RefreshCategory,
-    initialData: [] as CoreDNSClusterStatus[],
-    demoData: getDemoCoreDNSStatus(),
-    fetcher: async () => {
-      let pods: PodInfo[]
-      if (cluster) {
-        const data = await fetchAPI<{ pods: PodInfo[] }>('pods', { cluster, namespace: 'kube-system' })
-        pods = (data.pods || []).map(p => ({ ...p, cluster }))
-      } else {
-        pods = await fetchFromAllClusters<PodInfo>('pods', 'pods', { namespace: 'kube-system' })
-      }
-
-      const corednsPods = pods.filter(p =>
-        p.name?.includes('coredns') || p.name?.includes('kube-dns')
-      )
-
-      const byCluster = new Map<string, PodInfo[]>()
-      for (const pod of (corednsPods || [])) {
-        const c = pod.cluster || 'unknown'
-        if (!byCluster.has(c)) byCluster.set(c, [])
-        byCluster.get(c)!.push(pod)
-      }
-
-      const clusters = Array.from(byCluster.entries()).map(([clusterName, clusterPods]) => {
-        const running = clusterPods.filter(p => p.status === 'Running')
-        const healthy = running.length === clusterPods.length && clusterPods.length > 0
-        const totalRestarts = clusterPods.reduce((s, p) => s + (p.restarts || 0), 0)
-
-        return {
-          cluster: clusterName,
-          pods: clusterPods.map(p => ({
-            name: p.name,
-            status: p.status,
-            ready: p.ready,
-            restarts: p.restarts || 0,
-            version: p.containers?.[0]?.image?.split(':')[1]?.replace(/^v/, '') || '' })),
-          healthy,
-          totalRestarts } satisfies CoreDNSClusterStatus
-      })
-
-      // Sort clusters alphabetically for stable UI ordering
-      return clusters.sort((a, b) => a.cluster.localeCompare(b.cluster))
-    } })
-
-  return {
-    clusters: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-// ============================================================================
-// Additional Demo Data (for new cached hooks)
-// ============================================================================
-
-const getDemoGPUNodes = (): GPUNode[] => [
-  { name: 'gpu-node-1', cluster: 'vllm-gpu-cluster', gpuType: 'NVIDIA A100', gpuCount: 8, gpuAllocated: 5, acceleratorType: 'GPU', gpuMemoryMB: 81920, manufacturer: 'NVIDIA' },
-  { name: 'gpu-node-2', cluster: 'eks-prod-us-east-1', gpuType: 'NVIDIA T4', gpuCount: 4, gpuAllocated: 3, acceleratorType: 'GPU', gpuMemoryMB: 16384, manufacturer: 'NVIDIA' },
-  { name: 'gpu-node-3', cluster: 'gke-staging', gpuType: 'NVIDIA L4', gpuCount: 2, gpuAllocated: 0, acceleratorType: 'GPU', gpuMemoryMB: 24576, manufacturer: 'NVIDIA' },
-]
-
-const getDemoPVCs = (): PVC[] => [
-  { name: 'data-postgres-0', namespace: 'production', cluster: 'eks-prod-us-east-1', status: 'Bound', storageClass: 'gp3', capacity: '100Gi', accessModes: ['ReadWriteOnce'], age: '30d' },
-  { name: 'redis-data-0', namespace: 'data', cluster: 'gke-staging', status: 'Bound', storageClass: 'standard', capacity: '50Gi', accessModes: ['ReadWriteOnce'], age: '14d' },
-  { name: 'ml-scratch', namespace: 'ml-workloads', cluster: 'vllm-gpu-cluster', status: 'Pending', storageClass: 'fast-nvme', capacity: '500Gi', accessModes: ['ReadWriteMany'], age: '1h' },
-]
-
-const getDemoNamespaces = (): string[] =>
-  ['default', 'kube-system', 'kube-public', 'monitoring', 'production', 'staging', 'batch', 'data', 'ingress', 'security']
-
-const getDemoJobs = (): Job[] => [
-  { name: 'data-migration-v2', namespace: 'batch', cluster: 'eks-prod-us-east-1', status: 'Complete', completions: '1/1', duration: '5m', age: '2h' },
-  { name: 'model-training-run-42', namespace: 'ml-workloads', cluster: 'vllm-gpu-cluster', status: 'Running', completions: '0/1', age: '30m' },
-  { name: 'backup-db-daily', namespace: 'production', cluster: 'gke-staging', status: 'Failed', completions: '0/1', duration: '10m', age: '6h' },
-]
-
-const getDemoHPAs = (): HPA[] => [
-  { name: 'web-frontend', namespace: 'production', cluster: 'eks-prod-us-east-1', reference: 'Deployment/web-frontend', minReplicas: 2, maxReplicas: 10, currentReplicas: 4, targetCPU: '70%', currentCPU: '55%', age: '30d' },
-  { name: 'api-gateway', namespace: 'production', cluster: 'eks-prod-us-east-1', reference: 'Deployment/api-gateway', minReplicas: 3, maxReplicas: 20, currentReplicas: 8, targetCPU: '60%', currentCPU: '78%', age: '14d' },
-]
-
-const getDemoConfigMaps = (): ConfigMap[] => [
-  { name: 'app-config', namespace: 'production', cluster: 'eks-prod-us-east-1', dataCount: 5, age: '7d' },
-  { name: 'nginx-config', namespace: 'ingress', cluster: 'gke-staging', dataCount: 3, age: '14d' },
-]
-
-const getDemoSecrets = (): Secret[] => [
-  { name: 'db-credentials', namespace: 'production', cluster: 'eks-prod-us-east-1', type: 'Opaque', dataCount: 3, age: '30d' },
-  { name: 'tls-cert', namespace: 'ingress', cluster: 'eks-prod-us-east-1', type: 'kubernetes.io/tls', dataCount: 2, age: '90d' },
-]
-
-const getDemoServiceAccounts = (): ServiceAccount[] => [
-  { name: 'default', namespace: 'production', cluster: 'eks-prod-us-east-1', age: '90d' },
-  { name: 'prometheus', namespace: 'monitoring', cluster: 'gke-staging', age: '30d' },
-]
-
-const getDemoReplicaSets = (): ReplicaSet[] => [
-  { name: 'web-frontend-7d8f9c4b5', namespace: 'production', cluster: 'eks-prod-us-east-1', replicas: 3, readyReplicas: 3, ownerName: 'web-frontend', ownerKind: 'Deployment', age: '2d' },
-  { name: 'api-gateway-6c8d7f5e4', namespace: 'production', cluster: 'eks-prod-us-east-1', replicas: 3, readyReplicas: 1, ownerName: 'api-gateway', ownerKind: 'Deployment', age: '1d' },
-]
-
-const getDemoStatefulSets = (): StatefulSet[] => [
-  { name: 'postgres', namespace: 'production', cluster: 'eks-prod-us-east-1', replicas: 3, readyReplicas: 3, status: 'running', image: 'postgres:15', age: '30d' },
-  { name: 'redis', namespace: 'data', cluster: 'gke-staging', replicas: 3, readyReplicas: 3, status: 'running', image: 'redis:7', age: '14d' },
-]
-
-const getDemoDaemonSets = (): DaemonSet[] => [
-  { name: 'node-exporter', namespace: 'monitoring', cluster: 'eks-prod-us-east-1', desiredScheduled: 5, currentScheduled: 5, ready: 5, status: 'running', age: '60d' },
-  { name: 'fluentd', namespace: 'logging', cluster: 'gke-staging', desiredScheduled: 3, currentScheduled: 3, ready: 3, status: 'running', age: '30d' },
-]
-
-const getDemoCronJobs = (): CronJob[] => [
-  { name: 'db-backup', namespace: 'production', cluster: 'eks-prod-us-east-1', schedule: '0 2 * * *', suspend: false, active: 0, lastSchedule: new Date(Date.now() - 8 * 3600000).toISOString(), age: '60d' },
-  { name: 'log-cleanup', namespace: 'monitoring', cluster: 'gke-staging', schedule: '0 0 * * 0', suspend: false, active: 0, lastSchedule: new Date(Date.now() - 48 * 3600000).toISOString(), age: '30d' },
-]
-
-const getDemoIngresses = (): Ingress[] => [
-  { name: 'main-ingress', namespace: 'production', cluster: 'eks-prod-us-east-1', class: 'nginx', hosts: ['app.example.com', 'api.example.com'], address: '10.0.0.100', age: '30d' },
-  { name: 'staging-ingress', namespace: 'staging', cluster: 'gke-staging', class: 'nginx', hosts: ['staging.example.com'], address: '10.0.1.50', age: '14d' },
-]
-
-const getDemoNetworkPolicies = (): NetworkPolicy[] => [
-  { name: 'deny-all', namespace: 'production', cluster: 'eks-prod-us-east-1', policyTypes: ['Ingress', 'Egress'], podSelector: '{}', age: '60d' },
-  { name: 'allow-web', namespace: 'production', cluster: 'eks-prod-us-east-1', policyTypes: ['Ingress'], podSelector: 'app=web', age: '30d' },
-]
-
-const getDemoHelmReleases = (): HelmRelease[] => [
-  { name: 'prometheus', namespace: 'monitoring', revision: '5', updated: new Date(Date.now() - 2 * 3600000).toISOString(), status: 'deployed', chart: 'prometheus-25.8.0', app_version: '2.48.1', cluster: 'eks-prod-us-east-1' },
-  { name: 'grafana', namespace: 'monitoring', revision: '3', updated: new Date(Date.now() - 5 * 3600000).toISOString(), status: 'deployed', chart: 'grafana-7.0.11', app_version: '10.2.3', cluster: 'eks-prod-us-east-1' },
-  { name: 'nginx-ingress', namespace: 'ingress', revision: '8', updated: new Date(Date.now() - 24 * 3600000).toISOString(), status: 'deployed', chart: 'ingress-nginx-4.8.3', app_version: '1.9.4', cluster: 'gke-staging' },
-  { name: 'api-gateway', namespace: 'production', revision: '6', updated: new Date(Date.now() - 1 * 3600000).toISOString(), status: 'failed', chart: 'api-gateway-2.1.0', app_version: '3.5.0', cluster: 'eks-prod-us-east-1' },
-]
-
-const getDemoHelmHistory = (): HelmHistoryEntry[] => [
-  { revision: 6, updated: new Date(Date.now() - 1 * 3600000).toISOString(), status: 'failed', chart: 'api-gateway-2.1.0', app_version: '3.5.0', description: 'Upgrade failed: container crashed' },
-  { revision: 5, updated: new Date(Date.now() - 2 * 3600000).toISOString(), status: 'deployed', chart: 'prometheus-25.8.0', app_version: '2.48.1', description: 'Upgrade complete' },
-  { revision: 4, updated: new Date(Date.now() - 24 * 3600000).toISOString(), status: 'superseded', chart: 'prometheus-25.7.0', app_version: '2.48.0', description: 'Upgrade complete' },
-]
-
-const getDemoHelmValues = (): Record<string, unknown> => ({
-  replicaCount: 2,
-  image: { repository: 'prom/prometheus', tag: 'v2.48.1', pullPolicy: 'IfNotPresent' },
-  service: { type: 'ClusterIP', port: 9090 },
-  resources: { limits: { cpu: '500m', memory: '512Mi' }, requests: { cpu: '200m', memory: '256Mi' } } })
-
-const getDemoOperators = (): Operator[] => [
-  { name: 'prometheus-operator', namespace: 'monitoring', version: '0.72.0', status: 'Succeeded', cluster: 'eks-prod-us-east-1' },
-  { name: 'cert-manager', namespace: 'cert-manager', version: '1.14.0', status: 'Succeeded', upgradeAvailable: '1.15.0', cluster: 'eks-prod-us-east-1' },
-  { name: 'gpu-operator', namespace: 'nvidia-gpu-operator', version: '23.9.1', status: 'Succeeded', cluster: 'vllm-gpu-cluster' },
-]
-
-const getDemoOperatorSubscriptions = (): OperatorSubscription[] => [
-  { name: 'prometheus-sub', namespace: 'monitoring', channel: 'stable', source: 'community-operators', installPlanApproval: 'Automatic', currentCSV: 'prometheusoperator.0.72.0', cluster: 'eks-prod-us-east-1' },
-  { name: 'cert-manager-sub', namespace: 'cert-manager', channel: 'stable', source: 'community-operators', installPlanApproval: 'Manual', currentCSV: 'cert-manager.v1.14.0', pendingUpgrade: 'cert-manager.v1.15.0', cluster: 'eks-prod-us-east-1' },
-]
-
-const getDemoGitOpsDrifts = (): GitOpsDrift[] => [
-  { resource: 'nginx-deployment', namespace: 'production', cluster: 'eks-prod-us-east-1', kind: 'Deployment', driftType: 'modified', gitVersion: 'abc1234', details: 'replicas changed from 3 to 5', severity: 'medium' },
-  { resource: 'redis-config', namespace: 'data', cluster: 'gke-staging', kind: 'ConfigMap', driftType: 'modified', gitVersion: 'def5678', details: 'maxmemory-policy changed', severity: 'low' },
-]
-
-const getDemoBuildpackImages = (): BuildpackImage[] => [
-  { name: 'api-service', namespace: 'production', builder: 'paketo-buildpacks/builder-jammy-base', image: 'registry.example.com/api-service:latest', status: 'succeeded', updated: new Date(Date.now() - 2 * 3600000).toISOString(), cluster: 'eks-prod-us-east-1' },
-  { name: 'web-app', namespace: 'staging', builder: 'paketo-buildpacks/builder-jammy-full', image: 'registry.example.com/web-app:v2.1', status: 'building', updated: new Date(Date.now() - 300000).toISOString(), cluster: 'gke-staging' },
-]
-
-const getDemoK8sRoles = (): K8sRole[] => [
-  { name: 'pod-reader', namespace: 'production', cluster: 'eks-prod-us-east-1', isCluster: false, ruleCount: 3 },
-  { name: 'admin', cluster: 'eks-prod-us-east-1', isCluster: true, ruleCount: 15 },
-]
-
-const getDemoK8sRoleBindings = (): K8sRoleBinding[] => [
-  { name: 'pod-reader-binding', namespace: 'production', cluster: 'eks-prod-us-east-1', isCluster: false, roleName: 'pod-reader', roleKind: 'Role', subjects: [{ kind: 'User', name: 'jane' }] },
-  { name: 'admin-binding', cluster: 'eks-prod-us-east-1', isCluster: true, roleName: 'admin', roleKind: 'ClusterRole', subjects: [{ kind: 'Group', name: 'admins' }] },
-]
-
-const getDemoK8sServiceAccountsRbac = (): K8sServiceAccountInfo[] => [
-  { name: 'default', namespace: 'production', cluster: 'eks-prod-us-east-1', createdAt: new Date(Date.now() - 90 * 86400000).toISOString() },
-  { name: 'prometheus', namespace: 'monitoring', cluster: 'eks-prod-us-east-1', roles: ['prometheus-reader'], createdAt: new Date(Date.now() - 30 * 86400000).toISOString() },
-]
-
-// ============================================================================
-// Additional Cached Data Hooks
-// ============================================================================
-
-/**
- * Hook for fetching GPU nodes with caching
- */
-export function useCachedGPUNodes(
-  cluster?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<GPUNode[]> & { nodes: GPUNode[] } {
-  const { category = 'gpu' } = options || {}
-  const key = `gpuNodes:${cluster || 'all'}`
-
-  // Partial-failure protection (#8080, #8081): on the multi-cluster fan-out
-  // path, a transient error for the ONE cluster that actually has GPUs
-  // combined with empty results from the others is indistinguishable from
-  // "no GPUs anywhere". Rather than silently overwrite the cache with [],
-  // throw so useCache takes the error path and preserves the previous
-  // inventory. This is opt-in and GPU-specific — other endpoints (pods,
-  // deployments, etc.) still return partial data on partial failure.
-  const gpuFetchOptions: FetchFromAllClustersOptions = { throwIfPartialFailureEmpty: true }
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as GPUNode[],
-    demoData: getDemoGPUNodes(),
-    fetcher: async () => {
-      if (cluster) {
-        const raw = await fetchAPI<unknown>('gpu-nodes', { cluster })
-        const data = validateArrayResponse<{ nodes: GPUNode[] }>(GPUNodesResponseSchema, raw, '/api/mcp/gpu-nodes', 'nodes')
-        return (data.nodes || []).map(n => ({ ...n, cluster }))
-      }
-      return await fetchFromAllClusters<GPUNode>(
-        'gpu-nodes',
-        'nodes',
-        undefined,
-        true,
-        undefined,
-        gpuFetchOptions,
-      )
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return await fetchViaSSE<GPUNode>('gpu-nodes', 'nodes', {}, onProgress, gpuFetchOptions)
-    } })
-
-  return {
-    nodes: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching all pods (no namespace filter) with caching.
- * Used by GPU cards that need all pods across clusters for allocation tracking.
- */
-export function useCachedAllPods(
-  cluster?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<PodInfo[]> & { pods: PodInfo[] } {
-  const { category = 'pods' } = options || {}
-  const key = `allPods:${cluster || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as PodInfo[],
-    demoData: getDemoPods(),
-    fetcher: async () => {
-      if (cluster) {
-        const raw = await fetchAPI<unknown>('pods', { cluster })
-        const data = validateArrayResponse<{ pods: PodInfo[] }>(PodsResponseSchema, raw, '/api/mcp/pods (allPods)', 'pods')
-        return (data.pods || []).map(p => ({ ...p, cluster }))
-      }
-      return await fetchFromAllClusters<PodInfo>('pods', 'pods')
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return await fetchViaSSE<PodInfo>('pods', 'pods', {}, onProgress)
-    } })
-
-  return {
-    pods: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching PVCs with caching
- */
-export function useCachedPVCs(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<PVC[]> & { pvcs: PVC[] } {
-  const { category = 'default' } = options || {}
-  const key = `pvcs:${cluster || 'all'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as PVC[],
-    demoData: getDemoPVCs(),
-    fetcher: async () => {
-      if (cluster) {
-        const data = await fetchAPI<{ pvcs: PVC[] }>('pvcs', { cluster, namespace })
-        return (data.pvcs || []).map(p => ({ ...p, cluster }))
-      }
-      return await fetchFromAllClusters<PVC>('pvcs', 'pvcs', { namespace })
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return await fetchViaSSE<PVC>('pvcs', 'pvcs', { namespace }, onProgress)
-    } })
-
-  return {
-    pvcs: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching namespaces with caching.
- * Returns a list of namespace names for a given cluster.
- */
-export function useCachedNamespaces(
-  cluster?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<string[]> & { namespaces: string[] } {
-  const { category = 'namespaces' } = options || {}
-  const key = `namespaces:${cluster || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as string[],
-    demoData: getDemoNamespaces(),
-    fetcher: async () => {
-      if (!cluster) return getDemoNamespaces()
-      // Use the dedicated /api/namespaces endpoint which returns namespace details
-      const token = getToken()
-      if (!token) throw new Error('No authentication token')
-      const response = await fetch(`/api/namespaces?cluster=${encodeURIComponent(cluster)}`, {
-        headers: { 'Content-Type': 'application/json', Authorization: `Bearer ${token}` },
-        signal: AbortSignal.timeout(FETCH_DEFAULT_TIMEOUT_MS) })
-      if (!response.ok) throw new Error(`API error: ${response.status}`)
-      const data = await response.json().catch(() => null) as Array<{ name?: string; Name?: string }> | null
-      return (data || []).map((ns: { name?: string; Name?: string }) => ns.name || ns.Name || '').filter(Boolean)
-    } })
-
-  return {
-    namespaces: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching jobs with caching
- */
-export function useCachedJobs(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<Job[]> & { jobs: Job[] } {
-  const { category = 'default' } = options || {}
-  const key = `jobs:${cluster || 'all'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as Job[],
-    demoData: getDemoJobs(),
-    fetcher: async () => {
-      if (cluster) {
-        const data = await fetchAPI<{ jobs: Job[] }>('jobs', { cluster, namespace })
-        return (data.jobs || []).map(j => ({ ...j, cluster }))
-      }
-      return await fetchFromAllClusters<Job>('jobs', 'jobs', { namespace })
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return await fetchViaSSE<Job>('jobs', 'jobs', { namespace }, onProgress)
-    } })
-
-  return {
-    jobs: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching HPAs with caching
- */
-export function useCachedHPAs(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<HPA[]> & { hpas: HPA[] } {
-  const { category = 'default' } = options || {}
-  const key = `hpas:${cluster || 'all'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as HPA[],
-    demoData: getDemoHPAs(),
-    fetcher: async () => {
-      if (cluster) {
-        const data = await fetchAPI<{ hpas: HPA[] }>('hpas', { cluster, namespace })
-        return (data.hpas || []).map(h => ({ ...h, cluster }))
-      }
-      return await fetchFromAllClusters<HPA>('hpas', 'hpas', { namespace })
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return await fetchViaSSE<HPA>('hpas', 'hpas', { namespace }, onProgress)
-    } })
-
-  return {
-    hpas: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching ConfigMaps with caching
- */
-export function useCachedConfigMaps(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<ConfigMap[]> & { configmaps: ConfigMap[] } {
-  const { category = 'default' } = options || {}
-  const key = `configMaps:${cluster || 'all'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as ConfigMap[],
-    demoData: getDemoConfigMaps(),
-    fetcher: async () => {
-      if (cluster) {
-        const data = await fetchAPI<{ configmaps: ConfigMap[] }>('configmaps', { cluster, namespace })
-        return (data.configmaps || []).map(c => ({ ...c, cluster }))
-      }
-      return await fetchFromAllClusters<ConfigMap>('configmaps', 'configmaps', { namespace })
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return await fetchViaSSE<ConfigMap>('configmaps', 'configmaps', { namespace }, onProgress)
-    } })
-
-  return {
-    configmaps: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching Secrets with caching
- */
-export function useCachedSecrets(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<Secret[]> & { secrets: Secret[] } {
-  const { category = 'default' } = options || {}
-  const key = `secrets:${cluster || 'all'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as Secret[],
-    demoData: getDemoSecrets(),
-    fetcher: async () => {
-      if (cluster) {
-        const data = await fetchAPI<{ secrets: Secret[] }>('secrets', { cluster, namespace })
-        return (data.secrets || []).map(s => ({ ...s, cluster }))
-      }
-      return await fetchFromAllClusters<Secret>('secrets', 'secrets', { namespace })
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return await fetchViaSSE<Secret>('secrets', 'secrets', { namespace }, onProgress)
-    } })
-
-  return {
-    secrets: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching ServiceAccounts with caching
- */
-export function useCachedServiceAccounts(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<ServiceAccount[]> & { serviceAccounts: ServiceAccount[] } {
-  const { category = 'default' } = options || {}
-  const key = `serviceAccounts:${cluster || 'all'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as ServiceAccount[],
-    demoData: getDemoServiceAccounts(),
-    fetcher: async () => {
-      if (cluster) {
-        const data = await fetchAPI<{ serviceaccounts: ServiceAccount[] }>('serviceaccounts', { cluster, namespace })
-        return (data.serviceaccounts || []).map(sa => ({ ...sa, cluster }))
-      }
-      return await fetchFromAllClusters<ServiceAccount>('serviceaccounts', 'serviceaccounts', { namespace })
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return await fetchViaSSE<ServiceAccount>('serviceaccounts', 'serviceaccounts', { namespace }, onProgress)
-    } })
-
-  return {
-    serviceAccounts: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching ReplicaSets with caching
- */
-export function useCachedReplicaSets(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<ReplicaSet[]> & { replicasets: ReplicaSet[] } {
-  const { category = 'default' } = options || {}
-  const key = `replicaSets:${cluster || 'all'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as ReplicaSet[],
-    demoData: getDemoReplicaSets(),
-    fetcher: async () => {
-      if (cluster) {
-        const data = await fetchAPI<{ replicasets: ReplicaSet[] }>('replicasets', { cluster, namespace })
-        return (data.replicasets || []).map(rs => ({ ...rs, cluster }))
-      }
-      return await fetchFromAllClusters<ReplicaSet>('replicasets', 'replicasets', { namespace })
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return await fetchViaSSE<ReplicaSet>('replicasets', 'replicasets', { namespace }, onProgress)
-    } })
-
-  return {
-    replicasets: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching StatefulSets with caching
- */
-export function useCachedStatefulSets(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<StatefulSet[]> & { statefulsets: StatefulSet[] } {
-  const { category = 'default' } = options || {}
-  const key = `statefulSets:${cluster || 'all'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as StatefulSet[],
-    demoData: getDemoStatefulSets(),
-    fetcher: async () => {
-      if (cluster) {
-        const data = await fetchAPI<{ statefulsets: StatefulSet[] }>('statefulsets', { cluster, namespace })
-        return (data.statefulsets || []).map(ss => ({ ...ss, cluster }))
-      }
-      return await fetchFromAllClusters<StatefulSet>('statefulsets', 'statefulsets', { namespace })
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return await fetchViaSSE<StatefulSet>('statefulsets', 'statefulsets', { namespace }, onProgress)
-    } })
-
-  return {
-    statefulsets: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching DaemonSets with caching
- */
-export function useCachedDaemonSets(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<DaemonSet[]> & { daemonsets: DaemonSet[] } {
-  const { category = 'default' } = options || {}
-  const key = `daemonSets:${cluster || 'all'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as DaemonSet[],
-    demoData: getDemoDaemonSets(),
-    fetcher: async () => {
-      if (cluster) {
-        const data = await fetchAPI<{ daemonsets: DaemonSet[] }>('daemonsets', { cluster, namespace })
-        return (data.daemonsets || []).map(ds => ({ ...ds, cluster }))
-      }
-      return await fetchFromAllClusters<DaemonSet>('daemonsets', 'daemonsets', { namespace })
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return await fetchViaSSE<DaemonSet>('daemonsets', 'daemonsets', { namespace }, onProgress)
-    } })
-
-  return {
-    daemonsets: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching CronJobs with caching
- */
-export function useCachedCronJobs(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<CronJob[]> & { cronjobs: CronJob[] } {
-  const { category = 'default' } = options || {}
-  const key = `cronJobs:${cluster || 'all'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as CronJob[],
-    demoData: getDemoCronJobs(),
-    fetcher: async () => {
-      if (cluster) {
-        const data = await fetchAPI<{ cronjobs: CronJob[] }>('cronjobs', { cluster, namespace })
-        return (data.cronjobs || []).map(cj => ({ ...cj, cluster }))
-      }
-      return await fetchFromAllClusters<CronJob>('cronjobs', 'cronjobs', { namespace })
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return await fetchViaSSE<CronJob>('cronjobs', 'cronjobs', { namespace }, onProgress)
-    } })
-
-  return {
-    cronjobs: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching Ingresses with caching
- */
-export function useCachedIngresses(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<Ingress[]> & { ingresses: Ingress[] } {
-  const { category = 'default' } = options || {}
-  const key = `ingresses:${cluster || 'all'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as Ingress[],
-    demoData: getDemoIngresses(),
-    fetcher: async () => {
-      if (cluster) {
-        const data = await fetchAPI<{ ingresses: Ingress[] }>('ingresses', { cluster, namespace })
-        return (data.ingresses || []).map(i => ({ ...i, cluster }))
-      }
-      return await fetchFromAllClusters<Ingress>('ingresses', 'ingresses', { namespace })
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return await fetchViaSSE<Ingress>('ingresses', 'ingresses', { namespace }, onProgress)
-    } })
-
-  return {
-    ingresses: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching NetworkPolicies with caching
- */
-export function useCachedNetworkPolicies(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<NetworkPolicy[]> & { networkpolicies: NetworkPolicy[] } {
-  const { category = 'default' } = options || {}
-  const key = `networkPolicies:${cluster || 'all'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as NetworkPolicy[],
-    demoData: getDemoNetworkPolicies(),
-    fetcher: async () => {
-      if (cluster) {
-        const data = await fetchAPI<{ networkpolicies: NetworkPolicy[] }>('networkpolicies', { cluster, namespace })
-        return (data.networkpolicies || []).map(np => ({ ...np, cluster }))
-      }
-      return await fetchFromAllClusters<NetworkPolicy>('networkpolicies', 'networkpolicies', { namespace })
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return await fetchViaSSE<NetworkPolicy>('networkpolicies', 'networkpolicies', { namespace }, onProgress)
-    } })
-
-  return {
-    networkpolicies: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching Helm releases with caching (GitOps SSE endpoint)
- */
-export function useCachedHelmReleases(
-  cluster?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<HelmRelease[]> & { releases: HelmRelease[] } {
-  const { category = 'helm' } = options || {}
-  const key = `helmReleases:${cluster || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as HelmRelease[],
-    demoData: getDemoHelmReleases(),
-    fetcher: async () => {
-      const data = await fetchGitOpsAPI<{ releases: HelmRelease[] }>('helm-releases', cluster ? { cluster } : undefined)
-      return data.releases || []
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return await fetchViaGitOpsSSE<HelmRelease>('helm-releases', 'releases', {}, onProgress)
-    } })
-
-  return {
-    releases: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching Helm release history with caching
- */
-export function useCachedHelmHistory(
-  cluster?: string,
-  release?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<HelmHistoryEntry[]> & { history: HelmHistoryEntry[] } {
-  const { category = 'helm' } = options || {}
-  const key = `helmHistory:${cluster || 'none'}:${release || 'none'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as HelmHistoryEntry[],
-    demoData: getDemoHelmHistory(),
-    enabled: !!(cluster && release),
-    fetcher: async () => {
-      const data = await fetchGitOpsAPI<{ history: HelmHistoryEntry[] }>('helm-history', { cluster, release, namespace })
-      return data.history || []
-    } })
-
-  return {
-    history: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching Helm release values with caching
- */
-export function useCachedHelmValues(
-  cluster?: string,
-  release?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<Record<string, unknown>> & { values: Record<string, unknown> } {
-  const { category = 'helm' } = options || {}
-  const key = `helmValues:${cluster || 'none'}:${release || 'none'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: {} as Record<string, unknown>,
-    demoData: getDemoHelmValues(),
-    enabled: !!(cluster && release),
-    fetcher: async () => {
-      const data = await fetchGitOpsAPI<{ values: Record<string, unknown> }>('helm-values', { cluster, release, namespace })
-      return data.values || {}
-    } })
-
-  return {
-    values: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching operators with caching (GitOps SSE endpoint)
- */
-export function useCachedOperators(
-  cluster?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<Operator[]> & { operators: Operator[] } {
-  const { category = 'operators' } = options || {}
-  const key = `operators:${cluster || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as Operator[],
-    demoData: getDemoOperators(),
-    fetcher: async () => {
-      const data = await fetchGitOpsAPI<{ operators: Operator[] }>('operators', cluster ? { cluster } : undefined)
-      return data.operators || []
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return await fetchViaGitOpsSSE<Operator>('operators', 'operators', {}, onProgress)
-    } })
-
-  return {
-    operators: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching operator subscriptions with caching (GitOps SSE endpoint)
- */
-export function useCachedOperatorSubscriptions(
-  cluster?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<OperatorSubscription[]> & { subscriptions: OperatorSubscription[] } {
-  const { category = 'operators' } = options || {}
-  const key = `operatorSubscriptions:${cluster || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as OperatorSubscription[],
-    demoData: getDemoOperatorSubscriptions(),
-    fetcher: async () => {
-      const data = await fetchGitOpsAPI<{ subscriptions: OperatorSubscription[] }>('operator-subscriptions', cluster ? { cluster } : undefined)
-      return data.subscriptions || []
-    },
-    progressiveFetcher: cluster ? undefined : async (onProgress) => {
-      return await fetchViaGitOpsSSE<OperatorSubscription>('operator-subscriptions', 'subscriptions', {}, onProgress)
-    } })
-
-  return {
-    subscriptions: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching GitOps drift data with caching
- */
-export function useCachedGitOpsDrifts(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<GitOpsDrift[]> & { drifts: GitOpsDrift[] } {
-  const { category = 'gitops' } = options || {}
-  const key = `gitopsDrifts:${cluster || 'all'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as GitOpsDrift[],
-    demoData: getDemoGitOpsDrifts(),
-    fetcher: async () => {
-      const data = await fetchGitOpsAPI<{ drifts: GitOpsDrift[] }>('drifts', { cluster, namespace })
-      return data.drifts || []
-    } })
-
-  return {
-    drifts: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching buildpack images with caching
- */
-export function useCachedBuildpackImages(
-  cluster?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<BuildpackImage[]> & { images: BuildpackImage[] } {
-  const { category = 'default' } = options || {}
-  const key = `buildpackImages:${cluster || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as BuildpackImage[],
-    demoData: getDemoBuildpackImages(),
-    fetcher: async () => {
-      try {
-        const data = await fetchGitOpsAPI<{ images: BuildpackImage[] }>('buildpack-images', cluster ? { cluster } : undefined)
-        return data.images || []
-      } catch (err) {
-        // When no buildpacks CRDs exist on any cluster, the API returns 404.
-        // Treat this as an empty result rather than an error so the card
-        // settles into its empty state instead of retrying indefinitely.
-        if (err instanceof Error && err.message.includes('404')) {
-          return []
-        }
-        throw err
-      }
-    } })
-
-  return {
-    images: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching K8s Roles with caching (RBAC endpoint)
- */
-export function useCachedK8sRoles(
-  cluster?: string,
-  namespace?: string,
-  options?: { includeSystem?: boolean; category?: RefreshCategory }
-): CachedHookResult<K8sRole[]> & { roles: K8sRole[] } {
-  const { includeSystem = false, category = 'rbac' } = options || {}
-  const key = `k8sRoles:${cluster || 'all'}:${namespace || 'all'}:${includeSystem}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as K8sRole[],
-    demoData: getDemoK8sRoles(),
-    fetcher: async () => {
-      const data = await fetchRbacAPI<{ roles: K8sRole[] }>('roles', { cluster, namespace, includeSystem })
-      return data.roles || []
-    } })
-
-  return {
-    roles: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching K8s RoleBindings with caching (RBAC endpoint)
- */
-export function useCachedK8sRoleBindings(
-  cluster?: string,
-  namespace?: string,
-  options?: { includeSystem?: boolean; category?: RefreshCategory }
-): CachedHookResult<K8sRoleBinding[]> & { bindings: K8sRoleBinding[] } {
-  const { includeSystem = false, category = 'rbac' } = options || {}
-  const key = `k8sRoleBindings:${cluster || 'all'}:${namespace || 'all'}:${includeSystem}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as K8sRoleBinding[],
-    demoData: getDemoK8sRoleBindings(),
-    fetcher: async () => {
-      const data = await fetchRbacAPI<{ bindings: K8sRoleBinding[] }>('bindings', { cluster, namespace, includeSystem })
-      return data.bindings || []
-    } })
-
-  return {
-    bindings: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-/**
- * Hook for fetching K8s ServiceAccounts with caching (RBAC endpoint)
- */
-export function useCachedK8sServiceAccounts(
-  cluster?: string,
-  namespace?: string,
-  options?: { category?: RefreshCategory }
-): CachedHookResult<K8sServiceAccountInfo[]> & { serviceAccounts: K8sServiceAccountInfo[] } {
-  const { category = 'rbac' } = options || {}
-  const key = `k8sServiceAccounts:${cluster || 'all'}:${namespace || 'all'}`
-
-  const result = useCache({
-    key,
-    category,
-    initialData: [] as K8sServiceAccountInfo[],
-    demoData: getDemoK8sServiceAccountsRbac(),
-    fetcher: async () => {
-      const data = await fetchRbacAPI<{ serviceAccounts: K8sServiceAccountInfo[] }>('service-accounts', { cluster, namespace })
-      return data.serviceAccounts || []
-    } })
-
-  return {
-    serviceAccounts: result.data,
-    data: result.data,
-    isLoading: result.isLoading,
-    isRefreshing: result.isRefreshing,
-    isDemoFallback: result.isDemoFallback,
-    error: result.error,
-    isFailed: result.isFailed,
-    consecutiveFailures: result.consecutiveFailures,
-    lastRefresh: result.lastRefresh,
-    refetch: result.refetch }
-}
-
-// ============================================================================
-// ISO 27001 Security Audit — re-exported from useCachedISO27001.ts
-// ============================================================================
-
-export * from './useCachedISO27001'
-
